refactor: decompose clinical-matrix.ts and slim symptom-chat route.ts
clinical-matrix.ts (4268 lines) → re-export barrel + 5 focused sub-files:
  src/lib/clinical/types.ts              (183 lines — interfaces + factories)
  src/lib/clinical/symptom-map.ts        (1170 lines — SYMPTOM_MAP)
  src/lib/clinical/disease-db.ts         (906 lines  — DISEASE_DB)
  src/lib/clinical/breed-modifiers.ts    (235 lines  — BREED_MODIFIERS)
  src/lib/clinical/follow-up-questions.ts (1774 lines — FOLLOW_UP_QUESTIONS)
All 14 consumer files import from @/lib/clinical-matrix unchanged.

symptom-chat/route.ts (3104 lines → 2509 lines, −595) via 3 new lib files:
  src/lib/symptom-chat/route-shadow-telemetry.ts  (398 lines — LiveUpdateTarget + 8 shadow fns)
  src/lib/symptom-chat/quick-start-extraction.ts  (140 lines — 2 exported quick-start parsers)
  src/lib/symptom-chat/request-utils.ts            (99 lines  — MAX_REQUEST_BYTES, readJsonBody, jsonError)

Zero behavior change: all extracted code is identical; barrel re-exports
preserve every public symbol; route.ts imports updated accordingly.
