+++ src/app/api/ai/symptom-chat/route.ts
@@ -75,7 +75,6 @@ import {
   describeShadowModeDecision,
   getShadowModeDecision,
 } from "@/lib/sidecar-observability";
-import { appendShadowTelemetrySnapshot } from "@/lib/shadow-telemetry-store";
 import {
   buildContradictionRecord,
   detectTextContradictions,
@@ -94,12 +93,10 @@ import {
   extractDeterministicAnswersForTurn,
   mergeTurnAnswers,
 } from "@/lib/symptom-chat/answer-extraction";
-import { getQuickStartSymptomAliases } from "@/lib/symptom-chat/quick-start-symptoms";
 import {
   clearPendingQuestion,
   getClarificationAttemptCount,
   getPendingQuestionId,
-  getQuestionAskedCount,
   markPendingQuestionClarificationAttempt,
   pruneAnsweredQuestionState,
 } from "@/lib/symptom-chat/pending-question-state";
@@ -131,14 +128,10 @@ import { decideRepeatLoopGuard } from "@/lib/symptom-chat/repeat-loop-guard";
 import {
   buildSecondOpinionEligibilityTrace,
   extractSecondOpinionPendingAnswer,
-  getPrimarySuccessShadowSamplingAttemptCount as resolvePrimarySuccessShadowSamplingAttemptCount,
   getSecondOpinionExtractorMode,
-  type SecondOpinionExtractorMode,
-  type SecondOpinionExtractionResult,
 } from "@/lib/symptom-chat/second-opinion-extractor";
 import {
   createModelBudgetState,
-  type ModelBudgetState,
 } from "@/lib/model-budget";
 import { getModelRoute } from "@/lib/model-router";
 import {
@@ -187,9 +180,29 @@ import {
 } from "@/lib/azure/telemetry";
 import {
   normalizeWebPubSubSessionId,
-  publishTriageLiveUpdate,
   type TriageLiveUpdateStatus,
 } from "@/lib/azure/web-pubsub";
+import {
+  scheduleTriageLiveUpdate,
+  persistChatShadowTelemetrySnapshot,
+  getLatestSecondOpinionTraceObservation,
+  getCurrentTurnClarificationAttemptCount,
+  getPrimarySuccessShadowSamplingAttemptCount,
+  getSecondOpinionAcceptanceOutcome,
+  recordSecondOpinionNotRequestedTrace,
+  recordSecondOpinionShadowSample,
+  type LiveUpdateTarget,
+} from "@/lib/symptom-chat/route-shadow-telemetry";
+import {
+  shouldSuppressTextOnlyQuickStartKeywordSymptoms,
+  buildTextOnlyQuickStartExtraction,
+} from "@/lib/symptom-chat/quick-start-extraction";
+import {
+  MAX_REQUEST_BYTES,
+  readJsonBody,
+  jsonError,
+  type BodyParseResult,
+} from "@/lib/symptom-chat/request-utils";
 
 // =============================================================================
 // HYBRID STATE MACHINE API — 4-Model NVIDIA NIM Pipeline
@@ -229,511 +242,8 @@ interface RequestBody {
   gateOverrideToken?: string;
 }
 
-type LiveUpdateTarget = {
-  action: "chat" | "generate_report";
-  sessionId: string;
-  userId: string;
-};
-
-async function scheduleTriageLiveUpdate(
-  target: LiveUpdateTarget | null,
-  status: TriageLiveUpdateStatus
-): Promise<void> {
-  if (!target) {
-    return;
-  }
-
-  try {
-    await publishTriageLiveUpdate({
-      action: target.action,
-      sessionId: target.sessionId,
-      status,
-      userId: target.userId,
-    });
-  } catch {
-    // Live status is an enhancement; never let Web PubSub affect triage output.
-  }
-}
-
-async function persistChatShadowTelemetrySnapshot({
-  serviceCalls = [],
-  shadowComparisons = [],
-}: {
-  serviceCalls?: SidecarObservation[];
-  shadowComparisons?: ShadowComparisonRecord[];
-}): Promise<boolean> {
-  const snapshot = {
-    generatedAt: new Date().toISOString(),
-    recentServiceCalls: serviceCalls,
-    recentShadowComparisons: shadowComparisons,
-    source: "chat" as const,
-  };
-
-  try {
-    const persisted = await appendShadowTelemetrySnapshot(snapshot);
-    return persisted !== false;
-  } catch (shadowTelemetryError) {
-    console.error(
-      "[ShadowTelemetry] Failed to persist chat telemetry:",
-      shadowTelemetryError
-    );
-    return false;
-  }
-}
-
-function getLatestSecondOpinionTraceObservation(
-  session: TriageSession
-): SidecarObservation | undefined {
-  return [...(session.case_memory?.service_observations ?? [])]
-    .reverse()
-    .find(
-      (observation) =>
-        observation.service === "async-review-service" &&
-        observation.stage === "second_opinion"
-    );
-}
-
-function getCurrentTurnClarificationAttemptCount(
-  session: TriageSession,
-  questionId: string
-): number {
-  // The current owner reply is counted before the repeat-loop path persists it.
-  return getClarificationAttemptCount(session, questionId) + 1;
-}
-
-function getPrimarySuccessShadowSamplingAttemptCount(
-  session: TriageSession,
-  questionId: string
-): number {
-  return resolvePrimarySuccessShadowSamplingAttemptCount({
-    previousClarificationAttempts: getClarificationAttemptCount(
-      session,
-      questionId
-    ),
-    questionAskedCount: getQuestionAskedCount(session, questionId),
-  });
-}
-
-function getSecondOpinionAcceptanceOutcome(
-  result: SecondOpinionExtractionResult
-): SecondOpinionTraceTelemetry["acceptance_outcome"] {
-  if (result.status === "accepted") {
-    return "accepted";
-  }
-  if (result.status === "rejected") {
-    return "rejected";
-  }
-  if (result.status === "failed") {
-    return "failed";
-  }
-  return undefined;
-}
-
-async function recordSecondOpinionNotRequestedTrace({
-  session,
-  mode,
-  pendingQuestionId,
-  ownerMessage,
-  primaryExtractionFailed,
-  deterministicResolved,
-  clarificationAttempts,
-  hadUnresolved,
-  pendingAfter,
-  budgetState,
-}: {
-  session: TriageSession;
-  mode: SecondOpinionExtractorMode;
-  pendingQuestionId: string;
-  ownerMessage: string;
-  primaryExtractionFailed: boolean;
-  deterministicResolved: boolean;
-  clarificationAttempts: number;
-  hadUnresolved: boolean;
-  pendingAfter: boolean;
-  budgetState: ModelBudgetState;
-}): Promise<TriageSession> {
-  if (mode === "off") {
-    return session;
-  }
-
-  const eligibilityTrace = buildSecondOpinionEligibilityTrace({
-    mode,
-    pendingQuestionId,
-    ownerMessage,
-    primaryExtractionFailed,
-    deterministicResolved,
-    clarificationAttempts,
-    repeatGuardAlreadyFired: false,
-    budgetState,
-  });
-
-  if (eligibilityTrace.request_outcome === "requested") {
-    return session;
-  }
-
-  const secondOpinionTrace: SecondOpinionTraceTelemetry = {
-    ...eligibilityTrace,
-    comparison_append_outcome: "not_applicable",
-    comparison_write_outcome: "not_applicable",
-    extractor_reason: eligibilityTrace.eligibility_reason,
-  };
-  let nextSession = recordConversationTelemetry(session, {
-    event: "second_opinion",
-    turn_count: session.case_memory?.turn_count ?? 0,
-    question_id: pendingQuestionId,
-    outcome: "second_opinion_skipped",
-    source: "second_opinion",
-    reason: eligibilityTrace.eligibility_reason,
-    pending_before: hadUnresolved,
-    pending_after: pendingAfter,
-    second_opinion_trace: secondOpinionTrace,
-  });
-
-  const latestSecondOpinionTrace =
-    getLatestSecondOpinionTraceObservation(nextSession);
-  if (!latestSecondOpinionTrace) {
-    return nextSession;
-  }
-
-  const chatTelemetryPersisted = await persistChatShadowTelemetrySnapshot({
-    serviceCalls: [latestSecondOpinionTrace],
-    shadowComparisons: [],
-  });
-  if (!chatTelemetryPersisted) {
-    nextSession = recordConversationTelemetry(nextSession, {
-      event: "second_opinion",
-      turn_count: nextSession.case_memory?.turn_count ?? 0,
-      question_id: pendingQuestionId,
-      outcome: "second_opinion_failed",
-      source: "second_opinion",
-      reason: "telemetry_write_failed",
-      pending_before: hadUnresolved,
-      pending_after: pendingAfter,
-      second_opinion_trace: {
-        ...secondOpinionTrace,
-        extractor_reason: "telemetry_write_failed",
-      },
-    });
-  }
-
-  return nextSession;
-}
-
-async function recordSecondOpinionShadowSample({
-  session,
-  pendingQuestionId,
-  ownerMessage,
-  primaryExtractionFailed,
-  clarificationAttempts,
-  knownSymptomsBeforeTurn,
-  primaryAnswerValue,
-  hadUnresolved,
-}: {
-  session: TriageSession;
-  pendingQuestionId: string;
-  ownerMessage: string;
-  primaryExtractionFailed: boolean;
-  clarificationAttempts: number;
-  knownSymptomsBeforeTurn: string[];
-  primaryAnswerValue: string | boolean | number;
-  hadUnresolved: boolean;
-}): Promise<TriageSession> {
-  const budgetState = createModelBudgetState(
-    ensureStructuredCaseMemory(session).model_budget_state
-  );
-  const eligibilityTrace = buildSecondOpinionEligibilityTrace({
-    mode: "shadow",
-    pendingQuestionId,
-    ownerMessage,
-    primaryExtractionFailed,
-    deterministicResolved: true,
-    clarificationAttempts,
-    repeatGuardAlreadyFired: false,
-    budgetState,
-    isShadowSampling: true,
-  });
-
-  const shadowResult = await extractSecondOpinionPendingAnswer({
-    mode: "shadow",
-    pendingQuestionId,
-    ownerMessage,
-    primaryExtractionFailed,
-    deterministicResolved: true,
-    clarificationAttempts,
-    knownSymptomsBeforeTurn,
-    budgetState,
-    isShadowSampling: true,
-  });
-
-  if (shadowResult.budgetState) {
-    session = {
-      ...session,
-      case_memory: {
-        ...ensureStructuredCaseMemory(session),
-        model_budget_state: shadowResult.budgetState,
-      },
-    };
-  }
-
-  let recordedShadowComparison: ShadowComparisonRecord | undefined;
-  let comparisonAppendOutcome: SecondOpinionTraceTelemetry["comparison_append_outcome"] =
-    "not_applicable";
-  let comparisonWriteOutcome: SecondOpinionTraceTelemetry["comparison_write_outcome"] =
-    "not_applicable";
-
-  if (shadowResult.status === "accepted") {
-    const disagreed =
-      String(primaryAnswerValue) === String(shadowResult.answer.answerValue)
-        ? 0
-        : 1;
-    const shadowComparison = describeShadowComparison(
-      "async-review-service",
-      "primary_extraction_succeeded",
-      "second_opinion_extractor",
-      `q=${pendingQuestionId}; shadow_answer_recorded=true; conf=${shadowResult.answer.confidence.toFixed(2)}; agreed=${disagreed === 0}`,
-      disagreed
-    );
-    session = appendShadowComparison(session, shadowComparison);
-    const recordedShadowComparisons =
-      session.case_memory?.shadow_comparisons ?? [];
-    recordedShadowComparison =
-      recordedShadowComparisons[recordedShadowComparisons.length - 1];
-    if (recordedShadowComparison) {
-      comparisonAppendOutcome = "comparison_appended";
-      comparisonWriteOutcome = "comparison_write_succeeded";
-    } else {
-      comparisonAppendOutcome = "comparison_append_failed";
-      comparisonWriteOutcome = "comparison_write_failed";
-    }
-  }
-
-  const telemetryOutcome =
-    shadowResult.status === "accepted"
-      ? "second_opinion_used"
-      : shadowResult.status === "failed"
-        ? "second_opinion_failed"
-        : shadowResult.status === "rejected"
-          ? "second_opinion_rejected"
-          : eligibilityTrace.request_outcome === "budget_exhausted"
-            ? "second_opinion_failed"
-            : "second_opinion_skipped";
-  const secondOpinionTrace: SecondOpinionTraceTelemetry = {
-    ...eligibilityTrace,
-    acceptance_outcome: getSecondOpinionAcceptanceOutcome(shadowResult),
-    comparison_append_outcome: comparisonAppendOutcome,
-    comparison_write_outcome: comparisonWriteOutcome,
-    extractor_reason:
-      shadowResult.status === "accepted"
-        ? undefined
-        : shadowResult.reason ?? eligibilityTrace.eligibility_reason,
-  };
-  const gateEvents =
-    shadowResult.status === "accepted"
-      ? (["second_opinion_used"] as const)
-      : shadowResult.status === "failed"
-        ? (["second_opinion_failed"] as const)
-        : shadowResult.status === "rejected"
-          ? (["second_opinion_rejected"] as const)
-          : ([] as const);
-
-  session = recordConversationTelemetry(session, {
-    event: "second_opinion",
-    turn_count: session.case_memory?.turn_count ?? 0,
-    question_id: pendingQuestionId,
-    outcome: telemetryOutcome,
-    source: "second_opinion",
-    reason:
-      shadowResult.status === "accepted"
-        ? undefined
-        : shadowResult.reason ?? eligibilityTrace.eligibility_reason,
-    model:
-      shadowResult.status === "skipped"
-        ? undefined
-        : getModelRoute("extraction").primaryModel,
-    pending_before: hadUnresolved,
-    pending_after: false,
-    second_opinion_trace: secondOpinionTrace,
-    gate_events: [...gateEvents],
-  });
-
-  const latestSecondOpinionTrace =
-    getLatestSecondOpinionTraceObservation(session);
-  if (!latestSecondOpinionTrace) {
-    return session;
-  }
-
-  const chatTelemetryPersisted = await persistChatShadowTelemetrySnapshot({
-    serviceCalls: [latestSecondOpinionTrace],
-    shadowComparisons: recordedShadowComparison
-      ? [recordedShadowComparison]
-      : [],
-  });
-  if (chatTelemetryPersisted) {
-    return session;
-  }
-
-  const telemetryPersistenceFailureReason = recordedShadowComparison
-    ? "comparison_write_failed"
-    : "telemetry_write_failed";
-  return recordConversationTelemetry(session, {
-    event: "second_opinion",
-    turn_count: session.case_memory?.turn_count ?? 0,
-    question_id: pendingQuestionId,
-    outcome: "second_opinion_failed",
-    source: "second_opinion",
-    reason: telemetryPersistenceFailureReason,
-    pending_before: hadUnresolved,
-    pending_after: false,
-    second_opinion_trace: {
-      ...secondOpinionTrace,
-      comparison_write_outcome: recordedShadowComparison
-        ? "comparison_write_failed"
-        : secondOpinionTrace.comparison_write_outcome,
-      extractor_reason: telemetryPersistenceFailureReason,
-    },
-    gate_events: recordedShadowComparison ? ["second_opinion_failed"] : [],
-  });
-}
-
 function humanizeEmergencySignal(signal: string): string {
   return signal.replace(/_/g, " ").replace(/\s+/g, " ").trim();
-}
-
-function normalizeQuickStartText(value: string): string {
-  return value
-    .toLowerCase()
-    .replace(/[^a-z0-9\s]/g, " ")
-    .replace(/\s+/g, " ")
-    .trim();
-}
-
-function hasNegatedQuickStartPrefix(prefix: string): boolean {
-  return /\b(?:no|not|never|without|don t|dont|doesn t|doesnt|didn t|didnt|isn t|isnt|wasn t|wasnt|hasn t|hasnt|haven t|havent)\b/.test(
-    prefix
-  );
-}
-
-function deniesQuickStartSuffixSymptom(prefix: string): boolean {
-  return /\b(?:don t|dont|do not|doesn t|doesnt|does not|didn t|didnt|did not)\s+(?:think|believe|suspect|feel|know)\b/.test(
-    prefix
-  );
-}
-
-function getNormalizedQuickStartSymptomTexts(symptom: string): string[] {
-  return Array.from(
-    new Set(
-      [symptom.replace(/_/g, " "), ...getQuickStartSymptomAliases(symptom)]
-        .map(normalizeQuickStartText)
-        .filter(Boolean)
-    )
-  );
-}
-
-function isChipShapedQuickStartPrefix(prefix: string): boolean {
-  if (!prefix) {
-    return false;
-  }
-
-  const words = prefix.split(/\s+/).filter(Boolean);
-  if (words.length > 3) {
-    return false;
-  }
-
-  return !hasNegatedQuickStartPrefix(prefix);
-}
-
-function isExactTextOnlyQuickStartMessage(
-  rawMessage: string,
-  keywordSymptoms: string[]
-): boolean {
-  if (keywordSymptoms.length !== 1) {
-    return false;
-  }
-
-  const normalizedRawMessage = normalizeQuickStartText(rawMessage);
-  const normalizedSymptomTexts = getNormalizedQuickStartSymptomTexts(
-    keywordSymptoms[0]
-  );
-
-  if (normalizedSymptomTexts.includes(normalizedRawMessage)) {
-    return true;
-  }
-
-  // Quick-start chips are rendered as owner text like "Milo has been vomiting".
-  // Treat that app-generated sentence as the same deterministic single-symptom
-  // start, without broadening richer owner descriptions into this fast path.
-  const chipSuffix = normalizedSymptomTexts
-    .map((symptomText) => ` has been ${symptomText}`)
-    .find((suffix) => normalizedRawMessage.endsWith(suffix));
-  if (!chipSuffix) {
-    return false;
-  }
-
-  return isChipShapedQuickStartPrefix(
-    normalizedRawMessage.slice(0, -chipSuffix.length).trim()
-  );
-}
-
-function findQuickStartChipSuffix(
-  normalizedRawMessage: string,
-  symptom: string
-): string | null {
-  return (
-    getNormalizedQuickStartSymptomTexts(symptom)
-      .map((symptomText) => ` has been ${symptomText}`)
-      .find((suffix) => normalizedRawMessage.endsWith(suffix)) ?? null
-  );
-}
-
-function shouldSuppressTextOnlyQuickStartKeywordSymptoms(
-  rawMessage: string,
-  keywordSymptoms: string[]
-): boolean {
-  if (keywordSymptoms.length !== 1) {
-    return false;
-  }
-
-  const normalizedRawMessage = normalizeQuickStartText(rawMessage);
-  const chipSuffix = findQuickStartChipSuffix(
-    normalizedRawMessage,
-    keywordSymptoms[0]
-  );
-  if (!chipSuffix) {
-    return false;
-  }
-
-  return deniesQuickStartSuffixSymptom(
-    normalizedRawMessage.slice(0, -chipSuffix.length).trim()
-  );
-}
-
-function buildTextOnlyQuickStartExtraction(
-  session: TriageSession,
-  rawMessage: string,
-  keywordSymptoms: string[],
-  hasImage: boolean
-): {
-  symptoms: string[];
-  answers: Record<string, string | boolean | number>;
-} | null {
-  const hasPendingQuestion =
-    Boolean(session.last_question_asked) || Boolean(getPendingQuestionId(session));
-  if (
-    hasImage ||
-    hasPendingQuestion ||
-    session.known_symptoms.length > 0 ||
-    session.answered_questions.length > 0 ||
-    Object.keys(session.extracted_answers).length > 0 ||
-    session.red_flags_triggered.length > 0 ||
-    !isExactTextOnlyQuickStartMessage(rawMessage, keywordSymptoms)
-  ) {
-    return null;
-  }
-
-  return {
-    symptoms: keywordSymptoms,
-    answers: {},
-  };
 }
 
 function collectDeterministicEmergencySignals(
@@ -1234,109 +744,6 @@ function buildDeterministicEmergencyMessage(
   const details = signalSummary ? ` (${signalSummary})` : "";
 
   return `Based on the symptoms you've shared${details}, ${petName} may be having a medical emergency. Please go to the nearest emergency veterinary hospital now. I have enough information to prepare an emergency summary for the vet while you're on the way.`;
-}
-
-// Cap the request body (which can carry a base64 image) so an oversized or
-// buggy client cannot force unbounded buffering/parsing on the busiest AI
-// route. 10 MB matches the image-bearing async-review route. Pattern mirrors
-// the capped reader in symptom-check/route.ts. Route-local (not exported) so
-// the file keeps exporting only POST + maxDuration.
-const MAX_REQUEST_BYTES = 10 * 1024 * 1024;
-
-type BodyParseResult<T> =
-  | { ok: true; value: T }
-  | { ok: false; response: Response };
-
-function jsonError(error: string, status: number, code: string) {
-  return NextResponse.json({ error, code }, { status });
-}
-
-function decodeUtf8(chunks: Uint8Array[], totalBytes: number) {
-  const merged = new Uint8Array(totalBytes);
-  let offset = 0;
-
-  for (const chunk of chunks) {
-    merged.set(chunk, offset);
-    offset += chunk.byteLength;
-  }
-
-  return new TextDecoder().decode(merged);
-}
-
-async function readJsonBody<T>(
-  request: Request,
-  maxBytes: number
-): Promise<BodyParseResult<T>> {
-  const declaredLength = Number(request.headers.get("content-length") ?? "");
-  if (Number.isFinite(declaredLength) && declaredLength > maxBytes) {
-    return {
-      ok: false,
-      response: jsonError("Request body too large", 413, "PAYLOAD_TOO_LARGE"),
-    };
-  }
-
-  const reader = request.body?.getReader();
-  if (!reader) {
-    return {
-      ok: false,
-      response: jsonError("Request body is required", 400, "INVALID_JSON"),
-    };
-  }
-
-  const chunks: Uint8Array[] = [];
-  let totalBytes = 0;
-
-  try {
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) {
-        break;
-      }
-      if (!value) {
-        continue;
-      }
-
-      totalBytes += value.byteLength;
-      if (totalBytes > maxBytes) {
-        try {
-          await reader.cancel();
-        } catch {}
-
-        return {
-          ok: false,
-          response: jsonError(
-            "Request body too large",
-            413,
-            "PAYLOAD_TOO_LARGE"
-          ),
-        };
-      }
-
-      chunks.push(value);
-    }
-  } catch {
-    return {
-      ok: false,
-      response: jsonError("Malformed JSON body", 400, "INVALID_JSON"),
-    };
-  }
-
-  const rawBody = decodeUtf8(chunks, totalBytes).trim();
-  if (!rawBody) {
-    return {
-      ok: false,
-      response: jsonError("Request body is required", 400, "INVALID_JSON"),
-    };
-  }
-
-  try {
-    return { ok: true, value: JSON.parse(rawBody) as T };
-  } catch {
-    return {
-      ok: false,
-      response: jsonError("Malformed JSON body", 400, "INVALID_JSON"),
-    };
-  }
 }
 
 function sanitizeUnsupportedClientRedFlags(
