+++ src/lib/clinical/breed-modifiers.ts
@@ -1,0 +1,236 @@
+import type { BreedModifiers } from "./types";
+// --- BREED RISK MULTIPLIERS ---
+
+export const BREED_MODIFIERS: Record<string, BreedModifiers> = {
+  "Golden Retriever": {
+    hip_dysplasia: 2.2,
+    ccl_rupture: 1.6,
+    ibd: 2.8,
+    foreign_body: 3.0,
+    allergic_dermatitis: 2.0,
+    hypothyroidism: 2.5,
+    bone_cancer: 2.0,
+    ear_infection_bacterial: 2.0,
+    heart_disease: 1.5,
+    hot_spots: 3.0,
+    skin_mass: 1.8,
+  },
+  "Labrador Retriever": {
+    hip_dysplasia: 2.0,
+    ccl_rupture: 1.8,
+    osteoarthritis: 1.8,
+    foreign_body: 3.5,
+    allergic_dermatitis: 1.8,
+    ear_infection_bacterial: 2.2,
+    obesity_related: 2.5,
+    hot_spots: 2.8,
+    skin_mass: 1.5,
+  },
+  "German Shepherd": {
+    hip_dysplasia: 2.5,
+    lumbosacral_disease: 2.8,
+    ibd: 2.0,
+    impa: 1.5,
+    pancreatitis: 1.8,
+    degenerative_myelopathy: 3.0,
+    perianal_fistula: 4.0,
+    wound_infection: 1.3,
+    autoimmune_skin: 2.0,
+  },
+  Bulldog: {
+    difficulty_breathing: 4.0,
+    allergic_dermatitis: 3.0,
+    hip_dysplasia: 2.8,
+    patellar_luxation: 2.0,
+    ear_infection_bacterial: 2.5,
+    cherry_eye: 3.0,
+    heat_stroke: 3.5,
+    hot_spots: 2.5,
+    wound_infection: 1.5,
+    autoimmune_skin: 1.8,
+  },
+  "French Bulldog": {
+    difficulty_breathing: 4.5,
+    ivdd: 3.0,
+    allergic_dermatitis: 3.0,
+    ear_infection_bacterial: 2.0,
+    heat_stroke: 4.0,
+    hot_spots: 2.0,
+    wound_infection: 1.8,
+    autoimmune_skin: 1.5,
+  },
+  Pug: {
+    difficulty_breathing: 4.2,
+    heat_stroke: 3.8,
+    eye_disorders: 2.0,
+    allergic_dermatitis: 1.8,
+  },
+  Dachshund: {
+    ivdd: 12.0,
+    patellar_luxation: 1.5,
+    diabetes: 2.0,
+    cushings_disease: 2.0,
+    obesity_related: 2.0,
+  },
+  "Miniature Schnauzer": {
+    pancreatitis: 3.5,
+    diabetes: 2.2,
+  },
+  Poodle: {
+    allergic_dermatitis: 2.0,
+    ear_infection_bacterial: 2.5,
+    addisons_disease: 3.0,
+    bloat: 1.8,
+    diabetes: 1.5,
+  },
+  Boxer: {
+    heart_disease: 3.0,
+    bone_cancer: 2.5,
+    allergic_dermatitis: 2.0,
+    bloat: 2.0,
+    mast_cell_tumor: 3.5,
+    skin_mass: 3.0,
+    hot_spots: 1.5,
+  },
+  "Great Dane": {
+    gdv: 5.0,
+    bone_cancer: 3.0,
+    heart_disease: 2.5,
+    hip_dysplasia: 2.0,
+    wobbler_syndrome: 3.0,
+  },
+  "Irish Wolfhound": {
+    gdv: 4.5,
+    heart_disease: 2.8,
+    bone_cancer: 2.0,
+  },
+  Rottweiler: {
+    ccl_rupture: 2.5,
+    bone_cancer: 3.0,
+    hip_dysplasia: 2.0,
+    heart_disease: 1.8,
+    pancreatitis: 1.5,
+  },
+  "Yorkshire Terrier": {
+    patellar_luxation: 3.0,
+    collapsing_trachea: 3.5,
+    hypoglycemia: 3.0,
+    liver_shunt: 3.0,
+    dental_disease: 2.5,
+  },
+  Beagle: {
+    ear_infection_bacterial: 2.5,
+    ibd: 1.8,
+    hypothyroidism: 2.0,
+    epilepsy: 2.5,
+    obesity_related: 2.0,
+  },
+  Chihuahua: {
+    patellar_luxation: 3.5,
+    collapsing_trachea: 2.5,
+    heart_disease: 2.0,
+    hypoglycemia: 3.0,
+    dental_disease: 3.0,
+  },
+  "Cavalier King Charles Spaniel": {
+    heart_disease: 5.0,
+    syringomyelia: 4.0,
+    ear_infection_bacterial: 2.0,
+    allergic_dermatitis: 1.5,
+  },
+  "Doberman Pinscher": {
+    heart_disease: 4.0,
+    wobbler_syndrome: 3.0,
+    hypothyroidism: 2.5,
+    von_willebrands: 3.0,
+  },
+  "Cocker Spaniel": {
+    ear_infection_bacterial: 4.0,
+    allergic_dermatitis: 2.5,
+    hypothyroidism: 2.0,
+    imha: 3.0,
+    glaucoma: 2.5,
+  },
+  "Bernese Mountain Dog": {
+    bone_cancer: 4.0,
+    hip_dysplasia: 2.0,
+    ccl_rupture: 1.5,
+    bloat: 2.0,
+    histiocytic_sarcoma: 5.0,
+  },
+  Newfoundland: {
+    heart_disease: 3.5,
+    hip_dysplasia: 2.2,
+    bloat: 2.0,
+  },
+  "Siberian Husky": {
+    hip_dysplasia: 1.5,
+    hypothyroidism: 2.0,
+    eye_disorders: 2.5,
+    zinc_responsive_dermatosis: 3.0,
+  },
+  "Shih Tzu": {
+    eye_disorders: 3.0,
+    ear_infection_bacterial: 2.0,
+    dental_disease: 2.5,
+    kidney_disease: 1.5,
+    allergic_dermatitis: 1.8,
+  },
+  "Pit Bull": {
+    allergic_dermatitis: 3.5,
+    hot_spots: 2.0,
+    wound_infection: 1.5,
+    skin_mass: 1.8,
+    ccl_rupture: 2.0,
+    hip_dysplasia: 1.5,
+    autoimmune_skin: 1.5,
+  },
+  "Australian Shepherd": {
+    hip_dysplasia: 1.8,
+    epilepsy: 2.0,
+    eye_disorders: 2.5,
+    allergic_dermatitis: 1.5,
+    autoimmune_skin: 1.5,
+  },
+  "Border Collie": {
+    hip_dysplasia: 1.5,
+    epilepsy: 2.5,
+    eye_disorders: 2.0,
+    allergic_dermatitis: 1.3,
+  },
+  Pomeranian: {
+    patellar_luxation: 3.0,
+    collapsing_trachea: 2.5,
+    dental_disease: 2.5,
+    alopecia: 2.0,
+    heart_disease: 1.5,
+  },
+  Maltese: {
+    patellar_luxation: 2.5,
+    dental_disease: 3.0,
+    liver_shunt: 2.5,
+    collapsing_trachea: 2.0,
+    allergic_dermatitis: 1.5,
+  },
+  "Shar Pei": {
+    allergic_dermatitis: 4.0,
+    hot_spots: 3.0,
+    wound_infection: 2.5,
+    autoimmune_skin: 2.5,
+    ear_infection_bacterial: 3.0,
+    eye_disorders: 2.5,
+  },
+  "West Highland White Terrier": {
+    allergic_dermatitis: 4.0,
+    hot_spots: 2.5,
+    autoimmune_skin: 2.0,
+    skin_mass: 1.5,
+    liver_disease: 2.0,
+  },
+  "Pembroke Welsh Corgi": {
+    ivdd: 4.0,
+    obesity_related: 2.2,
+    hip_dysplasia: 1.8,
+  },
+};
+
