+++ src/lib/clinical/disease-db.ts
@@ -1,0 +1,916 @@
+import type { DiseaseEntry } from "./types";
+import {
+  DEFAULT_AGE_MODIFIER,
+  makeDiseaseEntry,
+  makeDermDisease,
+  makeNeuroOrOrthoDisease,
+  makeOphthalmicDisease,
+  makeRespiratoryDisease,
+  makeSystemicDisease,
+} from "./types";
+// --- DISEASE DATABASE ---
+
+const SUPPLEMENTAL_DISEASES: Record<string, DiseaseEntry> = {
+  liver_disease: makeSystemicDisease("Liver Disease", "Canine Hepatopathy", "Hepatic disease can drive appetite loss, lethargy, vomiting, jaundice, or increased thirst.", "high", 0.08),
+  dental_disease: makeSystemicDisease("Dental Disease", "Periodontal Disease / Oral Pain", "Oral pain commonly presents as reduced appetite, drooling, halitosis, or chewing reluctance.", "moderate", 0.12),
+  tooth_root_abscess: makeSystemicDisease("Tooth Root Abscess", "Periapical Tooth Root Abscess", "Infected tooth roots can cause oral pain, facial swelling under the eye, bad breath, and chewing difficulty.", "high", 0.05),
+  pain_general: makeSystemicDisease("Generalized Pain", "Non-localized Pain Syndrome", "Pain can present as lethargy, trembling, reduced appetite, and behavior change even without an obvious injury.", "high", 0.08),
+  parasites: makeSystemicDisease("Parasitism", "Intestinal Parasitic Disease", "Intestinal parasites can cause diarrhea, weight loss, blood in stool, or a poor hair coat.", "moderate", 0.12, { puppy: 2.2, adult: 1.0, senior: 0.9 }),
+  colitis: makeSystemicDisease("Colitis", "Large Bowel Colitis", "Colitis often causes frequent small-volume stool, mucus, urgency, and bright red blood.", "moderate", 0.12),
+  food_allergy: makeDermDisease("Food Allergy", "Adverse Food Reaction", "Food allergy can cause itching, recurrent ear disease, GI upset, or year-round skin inflammation.", "moderate", 0.1),
+  ivdd: makeNeuroOrOrthoDisease("Intervertebral Disc Disease", "Intervertebral Disc Extrusion / Protrusion", "Disc disease can cause back pain, reluctance to move, limb weakness, or sudden neurologic deficits.", "high", 0.09, { puppy: 0.4, adult: 1.0, senior: 1.8 }),
+  infection: makeSystemicDisease("Systemic Infection", "Infectious Inflammatory Disease", "Fever, lethargy, poor appetite, and pain can reflect bacterial, viral, or inflammatory infection.", "high", 0.08),
+  anemia: makeSystemicDisease("Anemia", "Reduced Red Blood Cell Mass", "Anemia can cause lethargy, pale gums, weakness, rapid breathing, and collapse.", "high", 0.07),
+  hypothyroidism: makeSystemicDisease("Hypothyroidism", "Canine Hypothyroidism", "Low thyroid hormone commonly causes lethargy, weight gain, skin disease, and exercise intolerance.", "moderate", 0.08, { puppy: 0.2, adult: 1.0, senior: 1.6 }),
+  addisons_disease: makeSystemicDisease("Addison's Disease", "Hypoadrenocorticism", "Addison's disease can cause waxing and waning lethargy, GI upset, weakness, trembling, or collapse.", "high", 0.06),
+  pneumonia: makeRespiratoryDisease("Pneumonia", "Pneumonia", "Lower airway infection or inflammation can cause cough, fast breathing, fever, and respiratory distress.", "high", 0.08),
+  collapsing_trachea: makeRespiratoryDisease("Collapsing Trachea", "Tracheal Collapse", "A dry honking cough, exercise intolerance, and noisy breathing are classic for collapsing trachea.", "moderate", 0.08),
+  laryngeal_paralysis: makeRespiratoryDisease("Laryngeal Paralysis", "Laryngeal Paralysis", "Upper airway obstruction can cause noisy breathing, heat intolerance, and worsening respiratory effort.", "high", 0.06, { puppy: 0.2, adult: 0.8, senior: 1.8 }),
+  lung_cancer: makeRespiratoryDisease("Lung Cancer", "Primary or Metastatic Pulmonary Neoplasia", "Pulmonary tumors can cause chronic cough, breathing difficulty, weight loss, and exercise intolerance.", "high", 0.03, { puppy: 0.1, adult: 0.7, senior: 2.0 }),
+  heart_failure: makeRespiratoryDisease("Heart Failure", "Congestive Heart Failure", "Fluid buildup from heart disease can cause cough, fast breathing, intolerance to exercise, and weakness.", "emergency", 0.06, { puppy: 0.2, adult: 0.8, senior: 1.9 }),
+  pleural_effusion: makeRespiratoryDisease("Pleural Effusion", "Pleural Space Effusion", "Fluid around the lungs causes shallow breathing, orthopnea, and severe respiratory effort.", "emergency", 0.04),
+  allergic_reaction: makeRespiratoryDisease("Allergic Reaction", "Acute Hypersensitivity Reaction", "Acute allergy can cause facial swelling, hives, vomiting, and rapid-onset breathing difficulty.", "emergency", 0.05),
+  trauma_chest: makeRespiratoryDisease("Chest Trauma", "Thoracic Trauma", "Blunt or penetrating chest trauma can cause pain, shock, internal bleeding, or breathing difficulty.", "emergency", 0.03),
+  flea_allergy: makeDermDisease("Flea Allergy Dermatitis", "Flea Allergy Dermatitis", "Even a small flea burden can trigger severe itching, hair loss, and hot spots.", "moderate", 0.09),
+  ear_infection: makeDermDisease("Ear Infection", "Otitis Externa", "Ear inflammation can cause odor, discharge, head shaking, and scratching around the ears.", "moderate", 0.08),
+  mange: makeDermDisease("Mange", "Demodectic or Sarcoptic Mange", "Mites can cause hair loss, crusting, intense itch, and secondary infection.", "moderate", 0.07),
+  yeast_infection: makeDermDisease("Yeast Dermatitis", "Malassezia Dermatitis", "Yeast overgrowth often causes greasy skin, odor, redness, and itch in skin folds or paws.", "moderate", 0.08),
+  cushings_disease: makeSystemicDisease("Cushing's Disease", "Hyperadrenocorticism", "Cushing's disease causes increased thirst, panting, recurrent skin issues, and abdominal enlargement.", "moderate", 0.06, { puppy: 0.1, adult: 0.8, senior: 1.8 }),
+  pyometra: makeSystemicDisease("Pyometra", "Infected Uterus", "Pyometra in intact females can cause increased drinking, lethargy, abdominal enlargement, or collapse.", "emergency", 0.05, { puppy: 0.0, adult: 1.0, senior: 1.4 }),
+  hypercalcemia: makeSystemicDisease("Hypercalcemia", "Elevated Blood Calcium", "High calcium can cause increased drinking, weakness, reduced appetite, and constipation.", "high", 0.03),
+  hypoglycemia: makeSystemicDisease("Hypoglycemia", "Low Blood Sugar", "Low blood glucose can cause trembling, weakness, disorientation, seizures, or collapse.", "emergency", 0.05, { puppy: 1.8, adult: 1.0, senior: 0.9 }),
+  seizure_disorder: makeNeuroOrOrthoDisease("Seizure Disorder", "Epileptic or Reactive Seizure Disorder", "Seizure disorders can cause trembling, collapse, salivation, paddling, or post-ictal confusion.", "high", 0.06),
+  fever: makeSystemicDisease("Fever", "Pyrexia", "Fever can cause lethargy, shivering, reduced appetite, and warm ears or body temperature.", "moderate", 0.08),
+  anxiety: makeSystemicDisease("Anxiety", "Anxiety / Stress Response", "Stress can cause trembling, pacing, panting, and restlessness without primary organic disease.", "low", 0.07),
+  ascites: makeSystemicDisease("Ascites", "Abdominal Effusion", "Fluid in the abdomen causes progressive distension and can reflect heart, liver, or protein disorders.", "high", 0.05),
+  splenic_mass: makeSystemicDisease("Splenic Mass", "Splenic Mass / Hemangiosarcoma Risk", "A splenic mass can cause abdominal enlargement, weakness, collapse, or internal bleeding.", "emergency", 0.04, { puppy: 0.0, adult: 0.8, senior: 1.8 }),
+  pregnancy: makeSystemicDisease("Pregnancy", "Canine Pregnancy", "Normal or complicated pregnancy can enlarge the abdomen and change appetite or behavior.", "moderate", 0.03, { puppy: 0.0, adult: 1.0, senior: 0.2 }),
+  hemorrhagic_gastroenteritis: makeSystemicDisease("Hemorrhagic Gastroenteritis", "Acute Hemorrhagic Diarrhea Syndrome", "Sudden profuse bloody diarrhea with vomiting can rapidly dehydrate dogs and become critical.", "emergency", 0.05),
+  coagulopathy: makeSystemicDisease("Coagulopathy", "Bleeding Disorder / Coagulopathy", "Abnormal clotting can cause GI bleeding, bruising, pale gums, and weakness.", "emergency", 0.04),
+  gi_cancer: makeSystemicDisease("Gastrointestinal Cancer", "Gastrointestinal Neoplasia", "GI tumors can cause chronic vomiting, blood in stool, weight loss, and reduced appetite.", "high", 0.03, { puppy: 0.0, adult: 0.6, senior: 1.8 }),
+  conjunctivitis: makeOphthalmicDisease("Conjunctivitis", "Conjunctivitis", "Conjunctival inflammation causes redness, discharge, squinting, and periocular irritation.", "moderate", 0.09),
+  corneal_ulcer: makeOphthalmicDisease("Corneal Ulcer", "Corneal Ulceration", "Corneal ulcers are painful and often cause squinting, tearing, cloudiness, and discharge.", "high", 0.07),
+  dry_eye: makeOphthalmicDisease("Dry Eye", "Keratoconjunctivitis Sicca", "Reduced tear production causes thick discharge, redness, and recurrent eye irritation.", "moderate", 0.06),
+  glaucoma: makeOphthalmicDisease("Glaucoma", "Glaucoma", "High intraocular pressure causes severe pain, redness, cloudy cornea, and sudden vision loss.", "emergency", 0.04),
+  uveitis: makeOphthalmicDisease("Uveitis", "Uveitis", "Inflammation inside the eye causes pain, redness, squinting, and vision change.", "high", 0.04),
+  entropion: makeOphthalmicDisease("Entropion", "Entropion", "Inward-rolling eyelids cause chronic discharge, corneal trauma, and squinting.", "moderate", 0.05),
+  ear_infection_yeast: makeDermDisease("Yeast Ear Infection", "Malassezia Otitis Externa", "Yeasty otitis causes dark debris, strong odor, itch, and head shaking.", "moderate", 0.06),
+  ear_mites: makeDermDisease("Ear Mites", "Otodectic Otitis", "Ear mites cause dark debris, irritation, and intense head shaking or scratching.", "moderate", 0.04, { puppy: 1.8, adult: 1.0, senior: 0.8 }),
+  foreign_body_ear: makeDermDisease("Ear Foreign Body", "Foreign Body in Ear Canal", "Grass awns and other material can trigger sudden pain, head shaking, and unilateral discharge.", "high", 0.04),
+  aural_hematoma: makeDermDisease("Aural Hematoma", "Aural Hematoma", "Bleeding within the ear flap causes a swollen, warm, pillow-like ear.", "moderate", 0.05),
+  hyperthyroidism: makeSystemicDisease("Hyperthyroidism-like Weight Loss", "Weight Loss with Increased Metabolic Drive", "Although uncommon in dogs, marked metabolic weight loss can resemble hyperthyroid patterns.", "moderate", 0.02, { puppy: 0.0, adult: 0.4, senior: 1.3 }),
+  cancer: makeSystemicDisease("Cancer", "Systemic Neoplasia", "Cancer can cause weight loss, lethargy, reduced appetite, bleeding, or masses depending on the site.", "high", 0.06),
+  exocrine_pancreatic_insufficiency: makeSystemicDisease("Exocrine Pancreatic Insufficiency", "Exocrine Pancreatic Insufficiency", "EPI causes weight loss, ravenous appetite, poor stool quality, and malabsorption.", "moderate", 0.04),
+  degenerative_myelopathy: makeNeuroOrOrthoDisease("Degenerative Myelopathy", "Degenerative Myelopathy", "Progressive hind-end weakness and ataxia in older dogs can mimic orthopedic limping.", "high", 0.04, { puppy: 0.0, adult: 0.6, senior: 1.8 }),
+  wobbler_syndrome: makeNeuroOrOrthoDisease("Wobbler Syndrome", "Cervical Spondylomyelopathy", "Neck disease can cause a wobbly gait, weakness, neck pain, and limb deficits.", "high", 0.03, { puppy: 0.1, adult: 0.8, senior: 1.5 }),
+  obesity_related: makeSystemicDisease("Obesity-related Mobility Disease", "Obesity-associated Musculoskeletal Strain", "Excess body weight can worsen joint pain, exercise intolerance, and mobility complaints.", "moderate", 0.06),
+  histiocytic_sarcoma: makeSystemicDisease("Histiocytic Sarcoma", "Histiocytic Sarcoma", "Aggressive cancer in predisposed breeds can present as lameness, masses, lethargy, or weight loss.", "high", 0.02, { puppy: 0.0, adult: 0.7, senior: 1.7 }),
+  imha: makeSystemicDisease("Immune-Mediated Hemolytic Anemia", "Immune-Mediated Hemolytic Anemia", "IMHA can cause sudden lethargy, pale or yellow gums, rapid breathing, and collapse.", "emergency", 0.03),
+  liver_shunt: makeSystemicDisease("Liver Shunt", "Portosystemic Shunt", "Abnormal liver blood flow can cause poor growth, vomiting, neurologic signs, and poor appetite.", "high", 0.03, { puppy: 2.0, adult: 0.8, senior: 0.3 }),
+  difficulty_breathing: makeRespiratoryDisease("Upper Airway Obstructive Disease", "Brachycephalic / Upper Airway Obstruction", "Chronic upper airway obstruction causes noisy breathing, panting intolerance, and exertional distress.", "high", 0.06),
+  heat_stroke: makeRespiratoryDisease("Heat Stroke", "Heat Stroke / Hyperthermia", "Heat illness causes panting, collapse, brick-red or pale gums, vomiting, and shock.", "emergency", 0.04),
+  bloat: makeSystemicDisease("Gastric Dilatation", "Gastric Dilatation / Bloat", "Gastric distension can cause a tense abdomen, restlessness, and rapid progression to GDV.", "emergency", 0.04),
+  mast_cell_tumor: makeDermDisease("Mast Cell Tumor", "Mast Cell Tumor", "Skin or subcutaneous masses can wax and wane, itch, redden, or ulcerate.", "high", 0.04, { puppy: 0.0, adult: 0.8, senior: 1.6 }),
+  perianal_fistula: makeDermDisease("Perianal Fistula", "Perianal Fistula / Anal Furunculosis", "Painful draining tracts near the anus can cause licking, odor, bleeding, and reluctance to sit.", "high", 0.03),
+  alopecia: makeDermDisease("Alopecia", "Alopecia", "Hair loss without severe inflammation can reflect endocrine, allergic, or follicular disease.", "low", 0.05),
+  zinc_responsive_dermatosis: makeDermDisease("Zinc-Responsive Dermatosis", "Zinc-Responsive Dermatosis", "A crusting dermatosis affecting face, pads, and pressure points in predisposed breeds.", "moderate", 0.02),
+  cherry_eye: makeOphthalmicDisease("Cherry Eye", "Prolapsed Gland of the Third Eyelid", "A pink tissue mass at the inner corner of the eye is typical of cherry eye.", "moderate", 0.05, { puppy: 1.4, adult: 1.0, senior: 0.8 }),
+  eye_disorders: makeOphthalmicDisease("Inherited Eye Disorders", "Inherited Ocular Disorders", "Breed-related eye disease can present with discharge, irritation, vision change, or recurrent eye pain.", "moderate", 0.03),
+  epilepsy: makeNeuroOrOrthoDisease("Epilepsy", "Idiopathic Epilepsy", "Recurrent seizures or episodic tremoring can reflect inherited epilepsy in predisposed breeds.", "high", 0.05, { puppy: 0.8, adult: 1.0, senior: 0.7 }),
+  syringomyelia: makeNeuroOrOrthoDisease("Syringomyelia", "Syringomyelia", "Neuropathic pain can cause phantom scratching, neck pain, vocalization, and gait change.", "high", 0.03),
+  von_willebrands: makeSystemicDisease("von Willebrand Disease", "von Willebrand Disease", "Inherited clotting disorders can cause bruising, mucosal bleeding, and excessive bleeding with minor trauma.", "high", 0.03),
+
+  // --- VET-902: New diseases for expanded complaint families ---
+
+  urinary_stones: makeSystemicDisease("Urinary Stones", "Urolithiasis / Urinary Calculi", "Mineral stones in the bladder or urethra cause straining, blood in urine, and potential blockage.", "high", 0.08),
+  urinary_infection: makeSystemicDisease("Urinary Tract Infection", "Bacterial Cystitis", "Bacterial infection of the bladder causes frequent urination, accidents, straining, and blood.", "moderate", 0.12),
+  prostate_disease: makeSystemicDisease("Prostate Disease", "Prostatitis / Benign Prostatic Hyperplasia / Prostatic Neoplasia", "Prostate enlargement causes straining to urinate, ribbon-like stool, and hind limb weakness.", "high", 0.06, { puppy: 0.0, adult: 1.0, senior: 2.2 }),
+  bladder_cancer: makeSystemicDisease("Bladder Cancer", "Transitional Cell Carcinoma", "Malignant bladder tumor causes blood in urine, straining, and recurrent infections.", "high", 0.02, { puppy: 0.0, adult: 0.4, senior: 2.0 }),
+  urethral_obstruction: makeSystemicDisease("Urethral Obstruction", "Urethral Blockage", "Complete blockage of urine flow is a life-threatening emergency causing bladder rupture and kidney failure.", "emergency", 0.04),
+  babesiosis: makeDiseaseEntry({
+    name: "Babesiosis",
+    medicalTerm: "Canine Babesiosis",
+    description: "Tick-borne hemolytic protozoal disease can cause sudden weakness, pale gums, dark urine, fever, and collapse.",
+    urgency: "emergency",
+    baseProbability: 0.02,
+    keyDifferentiators: [
+      "Acute weakness with hemolysis signals such as pale gums or dark urine",
+      "Tick exposure or travel to endemic areas raises concern",
+      "Fever, jaundice, or collapse can develop quickly",
+    ],
+    typicalTests: [
+      "CBC with blood smear review for hemolysis or organisms",
+      "Tick-borne infectious disease PCR or serology",
+      "Chemistry panel and urinalysis to assess systemic impact",
+    ],
+    typicalHomeCare: [
+      "Keep the dog calm and transport for same-day emergency evaluation",
+      "Do not delay care if weakness, pale gums, or collapse are present",
+    ],
+  }),
+
+  cognitive_dysfunction: makeNeuroOrOrthoDisease("Cognitive Dysfunction Syndrome", "Canine Cognitive Dysfunction", "Age-related brain changes cause confusion, disorientation, sleep changes, and housetraining loss.", "moderate", 0.08, { puppy: 0.0, adult: 0.3, senior: 2.5 }),
+  brain_tumor: makeNeuroOrOrthoDisease("Brain Tumor", "Intracranial Neoplasia", "Brain tumors cause seizures, behavior changes, vision loss, circling, and head pressing.", "high", 0.02, { puppy: 0.0, adult: 0.5, senior: 2.0 }),
+  vision_loss: makeSystemicDisease("Vision Loss", "Blindness / Visual Impairment", "Vision loss can be sudden (SARD, glaucoma) or gradual (cataract, retinal degeneration).", "high", 0.05),
+  hearing_loss: makeSystemicDisease("Hearing Loss", "Deafness / Hearing Impairment", "Hearing loss can be age-related, infection-related, or congenital in some breeds.", "moderate", 0.04),
+
+  megaesophagus: makeSystemicDisease("Megaesophagus", "Esophageal Dilatation and Hypomotility", "Enlarged esophagus causes regurgitation of undigested food, weight loss, and aspiration risk.", "high", 0.03),
+  vascular_ring_anomaly: makeSystemicDisease("Vascular Ring Anomaly", "Persistent Right Aortic Arch", "Congenital blood vessel compresses esophagus causing regurgitation in young dogs.", "high", 0.02, { puppy: 2.0, adult: 0.3, senior: 0.1 }),
+  myasthenia_gravis: makeNeuroOrOrthoDisease("Myasthenia Gravis", "Acquired Myasthenia Gravis", "Autoimmune neuromuscular disorder causes weakness, regurgitation, and exercise intolerance.", "high", 0.02),
+  esophageal_foreign_body: makeSystemicDisease("Esophageal Foreign Body", "Esophageal Obstruction", "Object lodged in esophagus causes drooling, retching, and inability to swallow.", "emergency", 0.03),
+  hiatal_hernia: makeSystemicDisease("Hiatal Hernia", "Sliding Hiatal Hernia", "Stomach protrudes through diaphragm causing regurgitation, especially after eating.", "moderate", 0.02),
+
+  obstipation: makeSystemicDisease("Obstipation", "Severe Refractory Constipation", "Severe, unresponsive constipation requiring medical or surgical intervention.", "high", 0.04),
+  prostate_enlargement: makeSystemicDisease("Prostate Enlargement", "Prostatomegaly", "Enlarged prostate compresses colon causing constipation and urination difficulty.", "moderate", 0.06, { puppy: 0.0, adult: 1.0, senior: 2.0 }),
+  perineal_hernia: makeSystemicDisease("Perineal Hernia", "Perineal Herniation", "Pelvic diaphragm weakness allows abdominal contents into perineal region causing straining.", "high", 0.03, { puppy: 0.0, adult: 0.6, senior: 2.0 }),
+  pelvic_canal_stenosis: makeSystemicDisease("Pelvic Canal Stenosis", "Narrowed Pelvic Canal", "Narrowed pelvic canal from prior fracture causes chronic constipation.", "moderate", 0.02),
+
+  polymyositis: makeNeuroOrOrthoDisease("Polymyositis", "Inflammatory Myopathy", "Immune-mediated muscle inflammation causes stiffness, pain, and weakness.", "high", 0.02),
+  exercise_induced_collapse: makeNeuroOrOrthoDisease("Exercise-Induced Collapse", "Exercise-Induced Collapse (EIC)", "Genetic disorder in certain breeds causes collapse after intense exercise.", "moderate", 0.03),
+  myopathy: makeNeuroOrOrthoDisease("Myopathy", "Muscle Disease", "Primary muscle disorders cause weakness, stiffness, and exercise intolerance.", "moderate", 0.03),
+
+  nasal_infection: makeRespiratoryDisease("Nasal Infection", "Rhinitis / Sinusitis", "Nasal cavity infection causes discharge, sneezing, and congestion.", "moderate", 0.06),
+  nasal_tumor: makeRespiratoryDisease("Nasal Tumor", "Nasal Neoplasia", "Nasal cavity tumors cause one-sided bloody discharge, facial deformity, and sneezing.", "high", 0.02, { puppy: 0.0, adult: 0.5, senior: 2.2 }),
+  nasal_foreign_body: makeRespiratoryDisease("Nasal Foreign Body", "Nasal Cavity Foreign Body", "Plant material or object in nasal cavity causes sudden sneezing and discharge.", "moderate", 0.03),
+  aspergillosis: makeRespiratoryDisease("Nasal Aspergillosis", "Fungal Rhinitis", "Fungal infection of nasal cavity causes bloody discharge, pain, and depigmentation.", "high", 0.02),
+  allergic_rhinitis: makeRespiratoryDisease("Allergic Rhinitis", "Nasal Allergy", "Environmental allergy causes clear nasal discharge and sneezing.", "low", 0.04),
+
+  vaginal_hyperplasia: makeSystemicDisease("Vaginal Hyperplasia", "Vaginal Edema / Hyperplasia", "Estrogen-driven swelling of vaginal tissue during heat cycle.", "moderate", 0.03),
+  metritis: makeSystemicDisease("Metritis", "Infected Uterus Post-Partum", "Uterine infection after whelping causes fever, foul discharge, and lethargy.", "emergency", 0.03),
+  vaginal_tumor: makeSystemicDisease("Vaginal Tumor", "Vaginal Neoplasia", "Vaginal mass causes discharge, straining, and visible tissue mass.", "moderate", 0.02),
+
+  testicular_tumor: makeSystemicDisease("Testicular Tumor", "Testicular Neoplasia", "Testicular cancer causes swelling, hormone changes, and metastasis risk.", "moderate", 0.04, { puppy: 0.0, adult: 0.8, senior: 1.8 }),
+  prostatitis: makeSystemicDisease("Prostatitis", "Bacterial Prostate Infection", "Prostate infection causes fever, pain, and urination/defecation difficulty.", "high", 0.04),
+  benign_prostatic_hyperplasia: makeSystemicDisease("Benign Prostatic Hyperplasia", "BPH", "Non-cancerous prostate enlargement common in intact males causes stool and urine changes.", "moderate", 0.08, { puppy: 0.0, adult: 1.0, senior: 1.5 }),
+
+  lymphoma: makeSystemicDisease("Lymphoma", "Multicentric Lymphoma", "Cancer of lymph nodes causing widespread swelling, lethargy, and weight loss.", "high", 0.04),
+  lipoma: makeDermDisease("Lipoma", "Benign Fatty Tumor", "Harmless fatty lump common in middle-aged and older dogs.", "low", 0.1, { puppy: 0.0, adult: 0.8, senior: 2.0 }),
+  hemangiosarcoma: makeSystemicDisease("Hemangiosarcoma", "Malignant Blood Vessel Tumor", "Aggressive cancer of blood vessels, commonly in spleen or heart, causes internal bleeding.", "emergency", 0.02, { puppy: 0.0, adult: 0.6, senior: 2.2 }),
+
+  seborrhea: makeDermDisease("Seborrhea", "Seborrheic Dermatitis", "Greasy, scaly skin condition causing odor, flaking, and secondary infection.", "moderate", 0.05),
+  folliculitis: makeDermDisease("Folliculitis", "Hair Follicle Inflammation / Infection", "Inflamed hair follicles cause bumps, pustules, and hair loss.", "moderate", 0.06),
+  immune_deficiency: makeSystemicDisease("Immune Deficiency", "Immunodeficiency Syndrome", "Weakened immune system causes recurrent infections of skin, ears, and respiratory tract.", "high", 0.02),
+
+  cauda_equina_syndrome: makeNeuroOrOrthoDisease("Cauda Equina Syndrome", "Cauda Equina Compression", "Nerve compression at tail end of spine causes fecal/urinary incontinence and hind limb weakness.", "high", 0.02),
+  anal_sphincter_incompetence: makeSystemicDisease("Anal Sphincter Incompetence", "Fecal Incontinence", "Loss of anal sphincter control causes passive fecal leakage.", "moderate", 0.02),
+
+  parvovirus: makeSystemicDisease("Parvovirus", "Canine Parvovirus Enteritis", "Highly contagious viral disease causing severe bloody diarrhea, vomiting, and rapid dehydration.", "emergency", 0.1, { puppy: 3.0, adult: 0.5, senior: 0.3 }),
+
+  superficial_pyoderma: makeDermDisease("Superficial Pyoderma", "Bacterial Skin Infection", "Surface bacterial infection causes pustules, crusts, and hair loss.", "moderate", 0.08),
+  demodicosis: makeDermDisease("Demodicosis", "Demodectic Mange", "Mite overgrowth causes hair loss, redness, and secondary infection, often in young dogs.", "moderate", 0.05, { puppy: 2.0, adult: 0.8, senior: 0.7 }),
+
+  sudden_acquired_retinal_degeneration: makeOphthalmicDisease("SARD", "Sudden Acquired Retinal Degeneration", "Sudden irreversible blindness with normal-appearing eyes initially.", "high", 0.02, { puppy: 0.0, adult: 0.6, senior: 1.8 }),
+  cataract: makeOphthalmicDisease("Cataract", "Lens Opacity", "Cloudy lens causes progressive vision loss, often bilateral.", "moderate", 0.06, { puppy: 0.2, adult: 0.8, senior: 2.0 }),
+  optic_neuritis: makeOphthalmicDisease("Optic Neuritis", "Optic Nerve Inflammation", "Optic nerve inflammation causes sudden vision loss and painful eye movement.", "high", 0.02),
+  hypertension: makeSystemicDisease("Hypertension", "Systemic Hypertension", "High blood pressure can cause sudden blindness, kidney damage, and neurologic signs.", "high", 0.04, { puppy: 0.0, adult: 0.6, senior: 2.2 }),
+
+  vestibular_disease: makeNeuroOrOrthoDisease("Vestibular Disease", "Idiopathic Vestibular Syndrome", "Inner ear/balance disorder causes head tilt, circling, nystagmus, and falling.", "moderate", 0.06, { puppy: 0.2, adult: 0.8, senior: 2.0 }),
+  age_related_deafness: makeSystemicDisease("Age-Related Deafness", "Presbycusis", "Progressive hearing loss in senior dogs.", "low", 0.06, { puppy: 0.0, adult: 0.4, senior: 2.5 }),
+  ototoxicity: makeSystemicDisease("Ototoxicity", "Drug-Induced Ear Damage", "Certain medications damage inner ear causing deafness and balance problems.", "high", 0.02),
+
+  oral_tumor: makeSystemicDisease("Oral Tumor", "Oral Neoplasia", "Malignant or benign growths in the mouth causing bleeding, odor, and eating difficulty.", "high", 0.04, { puppy: 0.0, adult: 0.6, senior: 2.0 }),
+  epulis: makeSystemicDisease("Epulis", "Gingival Epulis", "Benign gum tumor causing mass effect, bleeding, and eating difficulty.", "moderate", 0.04),
+  melanoma: makeSystemicDisease("Oral Melanoma", "Malignant Melanoma", "Aggressive oral cancer causing bleeding, tooth loss, and metastasis.", "emergency", 0.03, { puppy: 0.0, adult: 0.5, senior: 2.2 }),
+  squamous_cell_carcinoma: makeSystemicDisease("Squamous Cell Carcinoma", "Oral SCC", "Oral cancer causing ulceration, bleeding, and bone destruction.", "high", 0.03),
+  foreign_body_mouth: makeSystemicDisease("Mouth Foreign Body", "Oral Foreign Body", "Stick, bone, or object lodged in mouth or palate causing pain and drooling.", "moderate", 0.04),
+
+  stomatitis: makeSystemicDisease("Stomatitis", "Chronic Gingivostomatitis", "Severe oral inflammation causing pain, drooling, bad breath, and eating difficulty.", "high", 0.03),
+
+  fading_puppy_syndrome: makeSystemicDisease("Fading Puppy Syndrome", "Neonatal Mortality Syndrome", "Puppies that fail to thrive, cry constantly, and decline despite nursing.", "emergency", 0.08, { puppy: 3.0, adult: 0.0, senior: 0.0 }),
+  congenital_defect: makeSystemicDisease("Congenital Defect", "Birth Defect", "Structural abnormality present from birth causes failure to thrive or specific organ dysfunction.", "high", 0.04, { puppy: 2.5, adult: 0.2, senior: 0.0 }),
+
+  dystocia: makeSystemicDisease("Dystocia", "Difficult Birth", "Inability to deliver puppies naturally requires emergency intervention.", "emergency", 0.04),
+  eclampsia: makeSystemicDisease("Eclampsia", "Postpartum Hypocalcemia", "Life-threatening low calcium in nursing dams causes tremors, seizures, and collapse.", "emergency", 0.03),
+
+  seroma: makeSystemicDisease("Seroma", "Post-Surgical Fluid Accumulation", "Fluid pocket at surgery site, usually benign but can become infected.", "moderate", 0.06),
+  dehiscence: makeSystemicDisease("Wound Dehiscence", "Surgical Incision Opening", "Opening of surgical incision exposes internal tissues and risks infection.", "emergency", 0.03),
+  surgical_complication: makeSystemicDisease("Surgical Complication", "Post-Operative Complication", "Unexpected issue after surgery including bleeding, infection, or organ dysfunction.", "high", 0.06),
+
+  sepsis: makeSystemicDisease("Sepsis", "Systemic Inflammatory Response to Infection", "Life-threatening body-wide response to infection causing organ failure and shock.", "emergency", 0.03),
+
+  fibrocartilaginous_embolism: makeNeuroOrOrthoDisease("Fibrocartilaginous Embolism", "FCE / Spinal Stroke", "Spinal cord blood vessel blockage causes sudden one-sided paralysis without pain.", "high", 0.02),
+
+  // --- BENIGN / LOW-URGENCY DIFFERENTIALS (clinical-matrix benign expansion) ---
+  // Genuinely self-limiting or non-progressive conditions in an otherwise bright,
+  // alert, well-hydrated dog. These let mild presentations resolve to "monitor at
+  // home". They do NOT replace red-flag logic — any red flag still overrides.
+  dietary_indiscretion: makeDiseaseEntry({
+    name: "Dietary Indiscretion (Mild GI Upset)",
+    medicalTerm: "Acute Self-Limiting Dietary Indiscretion",
+    description:
+      "Mild, short-lived stomach upset after eating something new, rich, or off — a single bout of vomiting or soft stool in a dog that stays bright, alert, and drinking.",
+    urgency: "low",
+    baseProbability: 0.08,
+    ageModifier: { puppy: 0.7, adult: 1.0, senior: 0.8 },
+    keyDifferentiators: [
+      "Recent diet change, scavenging, table scraps, or new treat",
+      "One or two episodes only, dog otherwise bright and playful",
+      "No blood, still eating or interested in food, drinking normally",
+      "Resolves within 24-48 hours without escalation",
+    ],
+    typicalTests: [
+      "Usually clinical — based on history and a normal physical exam",
+      "Fecal exam if soft stool persists beyond a couple of days",
+    ],
+    typicalHomeCare: [
+      "Offer a small amount of water and rest the gut for a few hours",
+      "Feed a bland, easily digestible meal in small portions once settled",
+      "Return to normal food gradually over 2-3 days",
+      "Seek vet care if vomiting/diarrhea persists >48h, blood appears, or the dog becomes lethargic or stops drinking",
+    ],
+  }),
+  contact_dermatitis: makeDermDisease(
+    "Mild Contact / Seasonal Itch",
+    "Mild Contact or Seasonal Irritant Dermatitis",
+    "Mild localized itch or pink skin from contact with grass, pollen, or an irritant — no broken skin, no infection, and no foul odor.",
+    "low",
+    0.1,
+    { puppy: 1.1, adult: 1.0, senior: 0.9 }
+  ),
+  muscle_soreness: makeNeuroOrOrthoDisease(
+    "Mild Overexertion Muscle Soreness",
+    "Self-Limiting Myalgia from Overexertion",
+    "Transient general stiffness or mild soreness the day after unusually hard play or a long hike, with full weight-bearing and no specific painful spot.",
+    "low",
+    0.08,
+    { puppy: 0.9, adult: 1.0, senior: 0.9 }
+  ),
+  situational_stress: makeSystemicDisease(
+    "Situational Stress Response",
+    "Transient Situational / Fear-Related Stress",
+    "Short-lived trembling, pacing, or clinginess tied to a clear trigger (thunderstorm, fireworks, vet visit, new environment) that settles once the trigger passes.",
+    "low",
+    0.07,
+    { puppy: 1.2, adult: 1.0, senior: 1.0 }
+  ),
+};
+
+export const DISEASE_DB: Record<string, DiseaseEntry> = {
+  ccl_rupture: {
+    name: "Cranial Cruciate Ligament (CCL) Rupture",
+    medical_term: "Cranial Cruciate Ligament Rupture - Partial or Complete",
+    description:
+      "Tear of the cranial cruciate ligament in the stifle joint, analogous to ACL in humans.",
+    base_probability: 0.15,
+    age_modifier: { puppy: 0.3, adult: 1.0, senior: 1.8 },
+    urgency: "high",
+    key_differentiators: [
+      "Sudden onset lameness after activity",
+      "Weight-bearing but favoring leg",
+      "Stifle joint instability on drawer test",
+      "Effusion in stifle joint",
+    ],
+    typical_tests: [
+      "Orthopedic examination with cranial drawer test and tibial compression test",
+      "Orthogonal radiographs of stifle (lateral and craniocaudal views)",
+      "Sedated palpation for subtle partial tears",
+    ],
+    typical_home_care: [
+      "Strict exercise restriction — leash walks only for elimination",
+      "Anti-inflammatory protocol as prescribed by vet",
+      "Cold therapy 15 min every 4-6 hours for first 48 hours",
+    ],
+  },
+  hip_dysplasia: {
+    name: "Hip Dysplasia with Secondary Osteoarthritis",
+    medical_term: "Coxofemoral Dysplasia with Secondary Degenerative Joint Disease",
+    description:
+      "Malformation of the hip joint leading to progressive arthritis. Often subclinical until triggered.",
+    base_probability: 0.12,
+    age_modifier: { puppy: 0.5, adult: 1.0, senior: 2.2 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Stiffness worse after rest, improves with movement",
+      "Pain on hip extension",
+      "Bilateral symptoms common",
+      "Bunny-hopping gait",
+    ],
+    typical_tests: [
+      "Extended hip radiographs (ventrodorsal hip-extended view)",
+      "PennHIP distraction index measurement",
+      "OFA hip evaluation",
+    ],
+    typical_home_care: [
+      "Weight management — maintain lean body condition score 4-5/9",
+      "Joint supplement: glucosamine 500mg + chondroitin 400mg daily",
+      "Low-impact exercise — swimming, controlled leash walks",
+    ],
+  },
+  soft_tissue_injury: {
+    name: "Soft Tissue Strain / Muscle Contusion",
+    medical_term: "Acute Musculotendinous Strain with Localized Inflammation",
+    description: "Muscle or tendon strain from sudden activity. Usually self-limiting.",
+    base_probability: 0.25,
+    age_modifier: { puppy: 0.8, adult: 1.0, senior: 0.7 },
+    urgency: "low",
+    key_differentiators: [
+      "Acute onset after vigorous activity",
+      "Localized pain and warmth",
+      "Resolves within 5-7 days with rest",
+      "No joint instability",
+    ],
+    typical_tests: [
+      "Physical examination with palpation",
+      "Radiographs to rule out fracture if warranted",
+    ],
+    typical_home_care: [
+      "Rest and restricted activity for 5-7 days",
+      "Cold therapy for acute phase (first 48 hours)",
+      "Gradual return to activity",
+    ],
+  },
+  iliopsoas_strain: {
+    name: "Iliopsoas Muscle Strain",
+    medical_term: "Iliopsoas Myotendinopathy",
+    description:
+      "Deep hip flexor injury common in athletic dogs during sudden acceleration/deceleration.",
+    base_probability: 0.08,
+    age_modifier: { puppy: 0.4, adult: 1.2, senior: 0.6 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Pain on hip extension and internal rotation",
+      "Often mistaken for hip joint pathology",
+      "Common in active/athletic dogs",
+      "Worsens with activity vs improves (opposite of OA)",
+    ],
+    typical_tests: [
+      "Specific iliopsoas palpation under sedation",
+      "Musculoskeletal ultrasound of iliopsoas",
+      "MRI for chronic cases",
+    ],
+    typical_home_care: [
+      "Strict rest for 2-4 weeks",
+      "Physical rehabilitation exercises when cleared",
+      "Controlled leash walks only",
+    ],
+  },
+  osteoarthritis: {
+    name: "Osteoarthritis",
+    medical_term: "Degenerative Joint Disease (DJD)",
+    description: "Progressive joint degeneration causing chronic pain and stiffness.",
+    base_probability: 0.1,
+    age_modifier: { puppy: 0.1, adult: 0.8, senior: 3.0 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Gradual onset over weeks/months",
+      "Morning stiffness",
+      "Worsens in cold/damp weather",
+      "Muscle wasting over affected limb",
+    ],
+    typical_tests: [
+      "Orthogonal joint radiographs",
+      "Joint fluid analysis if effusion present",
+    ],
+    typical_home_care: [
+      "Weight management",
+      "Omega-3 fatty acid supplementation (EPA/DHA)",
+      "Gentle daily exercise",
+      "Orthopedic bedding",
+    ],
+  },
+  impa: {
+    name: "Immune-Mediated Polyarthritis (IMPA)",
+    medical_term: "Immune-Mediated Polyarthritis",
+    description:
+      "Autoimmune condition causing inflammation in multiple joints simultaneously.",
+    base_probability: 0.03,
+    age_modifier: { puppy: 0.5, adult: 1.5, senior: 1.0 },
+    urgency: "high",
+    key_differentiators: [
+      "Multiple joints affected simultaneously",
+      "Fever, lethargy, appetite loss",
+      "Shifting leg lameness",
+      "Joint swelling and warmth in multiple locations",
+    ],
+    typical_tests: [
+      "Joint fluid analysis (arthrocentesis) from multiple joints",
+      "Complete blood count with differential",
+      "ANA titer and rheumatoid factor",
+      "Tick-borne disease panel (4Dx)",
+    ],
+    typical_home_care: [
+      "Rest and activity restriction",
+      "Monitor temperature twice daily",
+      "Track which joints appear affected",
+    ],
+  },
+  bone_cancer: {
+    name: "Osteosarcoma",
+    medical_term: "Appendicular Osteosarcoma",
+    description:
+      "Aggressive bone cancer most common in large/giant breeds. Usually affects long bones.",
+    base_probability: 0.02,
+    age_modifier: { puppy: 0.1, adult: 0.5, senior: 2.5 },
+    urgency: "high",
+    key_differentiators: [
+      "Progressive lameness not improving with rest",
+      "Firm swelling at bone site",
+      "Pain increasing over weeks",
+      "Large/giant breed predisposition",
+    ],
+    typical_tests: [
+      "Orthogonal radiographs of affected limb",
+      "Thoracic radiographs (3 views) for metastasis screening",
+      "Bone biopsy for definitive diagnosis",
+      "Alkaline phosphatase levels",
+    ],
+    typical_home_care: [
+      "Pain management as prescribed",
+      "Activity restriction to comfort level",
+      "Monitor for pathologic fracture risk",
+    ],
+  },
+  lumbosacral_disease: {
+    name: "Lumbosacral Stenosis",
+    medical_term: "Degenerative Lumbosacral Stenosis with L7-S1 Nerve Root Compression",
+    description: "Compression of nerves at the lumbosacral junction causing hindlimb pain.",
+    base_probability: 0.04,
+    age_modifier: { puppy: 0.1, adult: 0.6, senior: 2.0 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Pain on lumbosacral palpation and tail elevation",
+      "Difficulty rising",
+      "Hindlimb weakness bilateral",
+      "Fecal/urinary incontinence in advanced cases",
+    ],
+    typical_tests: [
+      "Lumbosacral radiographs",
+      "MRI of lumbosacral spine",
+      "Neurologic examination",
+    ],
+    typical_home_care: [
+      "Weight management",
+      "Avoid jumping and stairs",
+      "Ramp access for vehicles/furniture",
+    ],
+  },
+  patellar_luxation: {
+    name: "Patellar Luxation",
+    medical_term: "Medial or Lateral Patellar Luxation",
+    description: "Kneecap slips out of its groove, causing intermittent lameness.",
+    base_probability: 0.08,
+    age_modifier: { puppy: 1.5, adult: 1.0, senior: 0.8 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Intermittent 'skipping' gait",
+      "Sudden non-weight-bearing then spontaneous resolution",
+      "More common in small breeds",
+      "Palpable luxation on exam",
+    ],
+    typical_tests: [
+      "Orthopedic exam with patellar palpation",
+      "Stifle radiographs",
+    ],
+    typical_home_care: [
+      "Weight management",
+      "Moderate consistent exercise",
+      "Joint supplements",
+    ],
+  },
+  gastroenteritis: {
+    name: "Acute Gastroenteritis",
+    medical_term: "Acute Gastroenteritis — Dietary Indiscretion vs Infectious",
+    description: "Inflammation of stomach/intestines from dietary indiscretion or infection.",
+    base_probability: 0.3,
+    age_modifier: { puppy: 1.5, adult: 1.0, senior: 0.8 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Vomiting and/or diarrhea",
+      "Recent dietary change or garbage access",
+      "Self-limiting within 24-48 hours",
+      "No abdominal pain on palpation",
+    ],
+    typical_tests: [
+      "Fecal analysis (flotation + direct smear)",
+      "CBC and chemistry panel if persists >48h",
+      "Abdominal radiographs if foreign body suspected",
+    ],
+    typical_home_care: [
+      "Withhold food 12 hours, then bland diet (boiled chicken + rice, 2:1 ratio)",
+      "Small frequent meals — 1/4 portion 4x daily",
+      "Ensure hydration — offer Pedialyte or low-sodium broth",
+    ],
+  },
+  pancreatitis: {
+    name: "Acute Pancreatitis",
+    medical_term: "Acute Pancreatitis",
+    description: "Inflammation of the pancreas causing severe abdominal pain and vomiting.",
+    base_probability: 0.1,
+    age_modifier: { puppy: 0.3, adult: 1.0, senior: 1.8 },
+    urgency: "high",
+    key_differentiators: [
+      "Severe abdominal pain — prayer position",
+      "Persistent vomiting",
+      "Anorexia",
+      "History of fatty food or garbage ingestion",
+    ],
+    typical_tests: [
+      "Canine pancreatic lipase immunoreactivity (cPLI/Spec cPL)",
+      "Abdominal ultrasound",
+      "CBC and comprehensive chemistry panel",
+    ],
+    typical_home_care: [
+      "NPO (nothing by mouth) until vet evaluation",
+      "Low-fat diet long-term once recovered",
+      "Monitor for dehydration — check skin turgor and gum moisture",
+    ],
+  },
+  foreign_body: {
+    name: "Gastrointestinal Foreign Body",
+    medical_term: "Gastrointestinal Foreign Body Obstruction",
+    description: "Ingested object blocking the GI tract — surgical emergency if complete obstruction.",
+    base_probability: 0.08,
+    age_modifier: { puppy: 2.5, adult: 1.0, senior: 0.4 },
+    urgency: "emergency",
+    key_differentiators: [
+      "Persistent vomiting — especially projectile",
+      "History of chewing/ingesting objects",
+      "Progressive lethargy and anorexia",
+      "Abdominal pain on palpation",
+    ],
+    typical_tests: [
+      "Abdominal radiographs (right lateral + VD)",
+      "Abdominal ultrasound",
+      "Barium contrast study if radiographs inconclusive",
+    ],
+    typical_home_care: [
+      "DO NOT attempt to induce vomiting without veterinary guidance",
+      "NPO — nothing by mouth",
+      "Seek emergency care immediately if suspected",
+    ],
+  },
+  ibd: {
+    name: "Inflammatory Bowel Disease (IBD)",
+    medical_term: "Inflammatory Bowel Disease — Lymphoplasmacytic or Eosinophilic",
+    description: "Chronic inflammatory condition of the GI tract causing intermittent symptoms.",
+    base_probability: 0.06,
+    age_modifier: { puppy: 0.3, adult: 1.2, senior: 1.5 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Chronic intermittent vomiting/diarrhea (weeks to months)",
+      "Weight loss despite adequate food intake",
+      "Fluctuating appetite",
+      "Symptoms wax and wane",
+    ],
+    typical_tests: [
+      "GI endoscopy with mucosal biopsies",
+      "Cobalamin (B12) and folate levels",
+      "Fecal analysis to rule out parasites",
+      "Hypoallergenic diet trial (8-12 weeks)",
+    ],
+    typical_home_care: [
+      "Novel protein or hydrolyzed diet trial",
+      "Food diary tracking symptoms vs diet",
+      "Probiotic supplementation",
+    ],
+  },
+  gdv: {
+    name: "Gastric Dilatation-Volvulus (GDV/Bloat)",
+    medical_term: "Gastric Dilatation-Volvulus (GDV)",
+    description: "Life-threatening twisting of the stomach. Minutes matter.",
+    base_probability: 0.02,
+    age_modifier: { puppy: 0.2, adult: 1.0, senior: 1.5 },
+    urgency: "emergency",
+    key_differentiators: [
+      "Distended/bloated abdomen",
+      "Unproductive retching — trying to vomit but nothing comes up",
+      "Restlessness and pacing",
+      "Rapid decline — pale gums, weak pulse, collapse",
+    ],
+    typical_tests: [
+      "Right lateral abdominal radiograph — double bubble sign",
+      "Emergency stabilization — IV fluids, gastric decompression",
+    ],
+    typical_home_care: [
+      "THIS IS AN EMERGENCY — go to nearest ER vet NOW",
+      "Do not wait, do not try home treatment",
+      "Call ahead so ER can prepare for surgery",
+    ],
+  },
+  toxin_ingestion: {
+    name: "Toxin Ingestion / Poisoning",
+    medical_term: "Acute Toxicosis",
+    description: "Ingestion of toxic substance — treatment depends on toxin type and timing.",
+    base_probability: 0.05,
+    age_modifier: { puppy: 2.0, adult: 1.0, senior: 0.6 },
+    urgency: "emergency",
+    key_differentiators: [
+      "Known or suspected access to toxic substance",
+      "Acute onset vomiting, trembling, or neurologic signs",
+      "Chocolate, xylitol, grapes, rodenticides, medications",
+    ],
+    typical_tests: [
+      "Toxicology screening based on suspected agent",
+      "CBC, chemistry, coagulation panel",
+      "Activated charcoal if within ingestion window",
+    ],
+    typical_home_care: [
+      "Call ASPCA Poison Control: (888) 426-4435",
+      "Do NOT induce vomiting unless directed by vet/poison control",
+      "Bring any packaging of suspected toxin to the ER",
+    ],
+  },
+  kidney_disease: {
+    name: "Chronic Kidney Disease",
+    medical_term: "Chronic Kidney Disease (CKD) — IRIS Staging",
+    description: "Progressive loss of kidney function.",
+    base_probability: 0.04,
+    age_modifier: { puppy: 0.2, adult: 0.5, senior: 3.0 },
+    urgency: "high",
+    key_differentiators: [
+      "Increased water intake and urination",
+      "Weight loss and decreased appetite",
+      "Vomiting and lethargy in advanced stages",
+      "Bad breath (uremic halitosis)",
+    ],
+    typical_tests: [
+      "BUN, creatinine, SDMA",
+      "Urinalysis with urine specific gravity",
+      "Urine protein:creatinine ratio",
+      "Abdominal ultrasound — kidney size and architecture",
+    ],
+    typical_home_care: [
+      "Ensure constant access to fresh water",
+      "Renal-specific diet (reduced phosphorus, moderate protein)",
+      "Monitor water intake — measure daily",
+    ],
+  },
+  diabetes: {
+    name: "Diabetes Mellitus",
+    medical_term: "Diabetes Mellitus — Type I (Insulin-Dependent)",
+    description: "Inability to regulate blood sugar due to insufficient insulin production.",
+    base_probability: 0.03,
+    age_modifier: { puppy: 0.1, adult: 1.0, senior: 2.0 },
+    urgency: "high",
+    key_differentiators: [
+      "Increased thirst and urination (PU/PD)",
+      "Weight loss despite good appetite",
+      "Cataracts (diabetic dogs)",
+      "Recurrent urinary tract infections",
+    ],
+    typical_tests: [
+      "Fasting blood glucose",
+      "Fructosamine level",
+      "Urinalysis — glucosuria and ketonuria",
+      "Complete chemistry panel",
+    ],
+    typical_home_care: [
+      "Monitor water intake",
+      "Consistent feeding schedule — same time, same amount",
+      "Monitor for signs of hypoglycemia or ketoacidosis",
+    ],
+  },
+  allergic_dermatitis: {
+    name: "Atopic Dermatitis / Allergic Skin Disease",
+    medical_term: "Canine Atopic Dermatitis (CAD)",
+    description: "Allergic skin condition causing chronic itching. Environmental or food-triggered.",
+    base_probability: 0.15,
+    age_modifier: { puppy: 1.5, adult: 1.0, senior: 0.7 },
+    urgency: "low",
+    key_differentiators: [
+      "Chronic itching — face, feet, ears, armpits, groin",
+      "Recurrent ear infections",
+      "Seasonal pattern (environmental) or constant (food)",
+      "Skin redness and secondary infections",
+    ],
+    typical_tests: [
+      "Skin scraping and cytology",
+      "Intradermal allergy testing",
+      "Elimination diet trial (8-12 weeks)",
+    ],
+    typical_home_care: [
+      "Medicated baths (chlorhexidine or oatmeal-based) 2x weekly",
+      "Omega-3 fatty acid supplementation",
+      "Flea prevention — consistent monthly application",
+      "Food diary if food allergy suspected",
+    ],
+  },
+  ear_infection_bacterial: {
+    name: "Bacterial Ear Infection (Otitis Externa)",
+    medical_term: "Otitis Externa — Bacterial",
+    description: "Infection of the external ear canal. Common secondary to allergies or moisture.",
+    base_probability: 0.2,
+    age_modifier: { puppy: 0.8, adult: 1.0, senior: 1.2 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Ear odor — foul or yeasty",
+      "Brown/yellow/green discharge",
+      "Head shaking and ear scratching",
+      "Pain on ear manipulation",
+    ],
+    typical_tests: [
+      "Otoscopic examination",
+      "Ear cytology — stain and microscopy",
+      "Culture and sensitivity if recurrent",
+    ],
+    typical_home_care: [
+      "Do not clean ears until vet evaluates — may have ruptured eardrum",
+      "Prevent water entry — no swimming, careful bathing",
+      "Complete full course of prescribed medication",
+    ],
+  },
+  heart_disease: {
+    name: "Cardiac Disease",
+    medical_term:
+      "Degenerative Mitral Valve Disease (DMVD) or Dilated Cardiomyopathy (DCM)",
+    description: "Heart disease causing reduced cardiac output. Progressive.",
+    base_probability: 0.04,
+    age_modifier: { puppy: 0.2, adult: 0.5, senior: 2.5 },
+    urgency: "high",
+    key_differentiators: [
+      "Exercise intolerance — tires quickly",
+      "Coughing — especially at night or after exercise",
+      "Rapid breathing at rest (>30 breaths/min)",
+      "Fainting episodes (syncope)",
+    ],
+    typical_tests: [
+      "Thoracic radiographs — heart size (VHS)",
+      "Echocardiogram",
+      "ProBNP blood test",
+      "ECG/Holter monitoring",
+    ],
+    typical_home_care: [
+      "Monitor resting respiratory rate — count breaths for 15 sec × 4, normal <30/min",
+      "Restrict strenuous exercise",
+      "Low-sodium diet",
+    ],
+  },
+  kennel_cough: {
+    name: "Kennel Cough (Infectious Tracheobronchitis)",
+    medical_term: "Canine Infectious Respiratory Disease Complex (CIRDC)",
+    description: "Highly contagious respiratory infection. Usually self-limiting.",
+    base_probability: 0.15,
+    age_modifier: { puppy: 2.0, adult: 1.0, senior: 1.3 },
+    urgency: "low",
+    key_differentiators: [
+      "Honking/goose-like dry cough",
+      "Recent boarding, grooming, or dog park exposure",
+      "Gagging or retching after coughing",
+      "Usually eating/drinking normally",
+    ],
+    typical_tests: [
+      "Usually clinical diagnosis based on history",
+      "Respiratory PCR panel if severe or not improving",
+      "Thoracic radiographs if pneumonia suspected",
+    ],
+    typical_home_care: [
+      "Isolate from other dogs for 10-14 days",
+      "Use harness instead of collar — avoid tracheal pressure",
+      "Humidifier in resting area",
+      "Honey (1 tsp for medium dog) to soothe throat — NOT for puppies <1yr",
+    ],
+  },
+  wound_infection: {
+    name: "Wound Infection",
+    medical_term: "Infected Traumatic Wound with Secondary Bacterial Contamination",
+    description: "Open wound with bacterial infection causing inflammation, pain, discharge, and potential systemic illness.",
+    base_probability: 0.25,
+    age_modifier: { puppy: 1.2, adult: 1.0, senior: 1.5 },
+    urgency: "high",
+    key_differentiators: [
+      "Purulent discharge — yellow/green pus",
+      "Foul odor from wound site",
+      "Surrounding erythema and warmth extending beyond wound margins",
+      "Pain disproportionate to wound size",
+      "Fever, lethargy, or appetite loss in systemic cases",
+    ],
+    typical_tests: [
+      "Wound culture and sensitivity — aerobic and anaerobic",
+      "Cytology of wound exudate — stain for bacteria and inflammatory cells",
+      "CBC with differential — look for neutrophilia/left shift",
+      "Blood glucose if diabetic or immunocompromised",
+    ],
+    typical_home_care: [
+      "Keep wound clean — gentle saline flush (1 tsp salt per 1 cup warm water) 2-3x daily",
+      "Prevent licking — E-collar (cone) mandatory 24/7",
+      "Monitor for spreading redness — mark the border with a pen to track progression",
+      "Do NOT apply hydrogen peroxide or alcohol — these damage healing tissue",
+    ],
+  },
+  abscess: {
+    name: "Subcutaneous Abscess",
+    medical_term: "Subcutaneous Abscess with Localized Bacterial Infection",
+    description: "Pocket of pus under the skin, often from bite wounds or foreign body penetration.",
+    base_probability: 0.15,
+    age_modifier: { puppy: 0.8, adult: 1.0, senior: 1.2 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Firm, warm, painful swelling that may rupture and drain",
+      "Often preceded by a bite wound or puncture",
+      "Foul-smelling purulent drainage if ruptured",
+      "May cause fever and regional lymph node enlargement",
+    ],
+    typical_tests: [
+      "Fine needle aspirate of mass — cytology",
+      "Culture and sensitivity of aspirated material",
+      "Ultrasound if deep abscess suspected",
+    ],
+    typical_home_care: [
+      "Warm compress 10-15 min 3-4x daily to encourage drainage",
+      "Keep draining abscess clean with saline flush",
+      "E-collar to prevent licking",
+      "Complete full antibiotic course as prescribed",
+    ],
+  },
+  hot_spots: {
+    name: "Acute Moist Dermatitis (Hot Spot)",
+    medical_term: "Acute Moist Dermatitis — Pyotraumatic Dermatitis",
+    description: "Rapidly developing area of inflamed, infected skin from self-trauma (licking/scratching).",
+    base_probability: 0.2,
+    age_modifier: { puppy: 0.8, adult: 1.0, senior: 0.9 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Rapidly developing (hours) moist, red, painful lesion",
+      "Often under matted fur — may be larger than visible",
+      "Self-trauma driven — dog constantly licking/chewing the area",
+      "Common in thick-coated breeds especially in warm/humid weather",
+    ],
+    typical_tests: [
+      "Skin scraping to rule out demodex mites",
+      "Impression cytology — bacteria and yeast",
+      "Flea comb and skin inspection for underlying cause",
+    ],
+    typical_home_care: [
+      "Clip fur around the lesion — 1 inch margin for air exposure",
+      "Clean with dilute chlorhexidine solution 2x daily",
+      "E-collar is mandatory — self-trauma is the #1 reason they worsen",
+      "Keep area dry — no swimming, careful bathing",
+    ],
+  },
+  laceration: {
+    name: "Traumatic Laceration",
+    medical_term: "Traumatic Laceration with Possible Subcutaneous Tissue Involvement",
+    description: "Cut or tear in the skin from trauma. Depth determines whether sutures or surgery are needed.",
+    base_probability: 0.2,
+    age_modifier: { puppy: 1.3, adult: 1.0, senior: 0.8 },
+    urgency: "high",
+    key_differentiators: [
+      "Visible break in skin continuity",
+      "Bleeding (active or recent)",
+      "Visible subcutaneous fat or deeper tissues if full-thickness",
+      "Clean edges (sharp object) vs ragged edges (bite/tear)",
+    ],
+    typical_tests: [
+      "Wound exploration under sedation to assess depth",
+      "Radiographs if penetrating wound or foreign body suspected",
+      "Culture if wound is contaminated or old (>6 hours)",
+    ],
+    typical_home_care: [
+      "Apply direct pressure with clean cloth for active bleeding — hold 5+ minutes",
+      "Cover loosely with clean bandage — do NOT apply tight tourniquets",
+      "Wounds >1cm deep, >2cm long, or on joints need veterinary closure",
+      "Do NOT attempt to close or glue wounds at home",
+    ],
+  },
+  skin_mass: {
+    name: "Skin Mass / Tumor",
+    medical_term: "Cutaneous or Subcutaneous Neoplasia — Type Pending Histopathology",
+    description: "Lump or mass on/under the skin. Can be benign (lipoma) or malignant (mast cell tumor, soft tissue sarcoma).",
+    base_probability: 0.08,
+    age_modifier: { puppy: 0.2, adult: 0.6, senior: 2.5 },
+    urgency: "moderate",
+    key_differentiators: [
+      "New lump/bump that persists >2 weeks",
+      "Rapid growth, irregular borders, or color change",
+      "Fixed to underlying tissue (concerning) vs freely movable (often benign)",
+      "Ulceration or bleeding from mass surface",
+    ],
+    typical_tests: [
+      "Fine needle aspirate with cytology — first-line for ALL skin masses",
+      "Incisional or excisional biopsy with histopathology",
+      "Staging: thoracic radiographs, abdominal ultrasound, regional lymph node aspirate",
+    ],
+    typical_home_care: [
+      "Measure and photograph the mass weekly to track growth",
+      "ANY new mass should be aspirated by a vet — do not wait and watch",
+      "Prevent licking/scratching if ulcerated",
+    ],
+  },
+  autoimmune_skin: {
+    name: "Autoimmune Skin Disease",
+    medical_term: "Pemphigus Foliaceus or Discoid Lupus Erythematosus",
+    description: "Immune system attacks the skin causing crusting, ulceration, and depigmentation.",
+    base_probability: 0.02,
+    age_modifier: { puppy: 0.3, adult: 1.0, senior: 1.3 },
+    urgency: "moderate",
+    key_differentiators: [
+      "Crusting and ulceration on nose, ear tips, footpads",
+      "Symmetrical distribution",
+      "Depigmentation of nasal planum",
+      "Not responsive to antibiotics alone",
+    ],
+    typical_tests: [
+      "Skin biopsy — punch biopsy from lesion margin",
+      "Direct immunofluorescence",
+      "ANA titer",
+      "CBC and chemistry panel",
+    ],
+    typical_home_care: [
+      "Avoid sun exposure — UV worsens many autoimmune skin conditions",
+      "Gentle cleaning of crusted areas with saline",
+      "Do not attempt to pick or remove crusts forcefully",
+    ],
+  },
+  ...SUPPLEMENTAL_DISEASES,
+};
+
