+++ src/lib/clinical/symptom-map.ts
@@ -1,0 +1,1171 @@
+import type { SymptomEntry } from "./types";
+// --- SYMPTOM → DISEASE + QUESTION MAP ---
+
+export const SYMPTOM_MAP: Record<string, SymptomEntry> = {
+  vomiting: {
+    linked_diseases: [
+      "dietary_indiscretion",
+      "gastroenteritis",
+      "pancreatitis",
+      "foreign_body",
+      "ibd",
+      "gdv",
+      "toxin_ingestion",
+      "kidney_disease",
+      "coagulopathy",
+    ],
+    follow_up_questions: [
+      "vomit_duration",
+      "vomit_frequency",
+      "vomit_blood",
+      "vomit_content",
+      "toxin_exposure",
+      "dietary_change",
+      "appetite_status",
+      "spay_neuter_status",
+      "prior_similar_episode",
+      "recent_diet_change",
+      "last_vet_visit",
+    ],
+    red_flags: [
+      "vomit_blood",
+      "unproductive_retching",
+      "toxin_confirmed",
+      "not_drinking",
+      "pale_gums",
+    ],
+    body_systems: ["gastrointestinal"],
+  },
+  not_eating: {
+    linked_diseases: [
+      "gastroenteritis",
+      "pancreatitis",
+      "foreign_body",
+      "kidney_disease",
+      "liver_disease",
+      "dental_disease",
+      "pain_general",
+    ],
+    follow_up_questions: [
+      "appetite_duration",
+      "water_intake",
+      "weight_loss",
+      "treats_accepted",
+      "spay_neuter_status",
+    ],
+    red_flags: ["no_water_24h"],
+    body_systems: ["gastrointestinal", "systemic"],
+  },
+  diarrhea: {
+    linked_diseases: [
+      "dietary_indiscretion",
+      "gastroenteritis",
+      "pancreatitis",
+      "ibd",
+      "parasites",
+      "colitis",
+      "food_allergy",
+      "hemorrhagic_gastroenteritis",
+      "parvovirus",
+      "coagulopathy",
+    ],
+    follow_up_questions: [
+      "stool_blood",
+      "blood_amount",
+      "stool_frequency",
+      "stool_consistency",
+      "diarrhea_duration",
+      "gum_color",
+      "water_intake",
+      "dietary_change",
+    ],
+    red_flags: [
+      "stool_blood_large",
+      "bloody_diarrhea_puppy",
+      "large_blood_volume",
+      "pale_gums",
+    ],
+    body_systems: ["gastrointestinal"],
+  },
+  limping: {
+    linked_diseases: [
+      "ccl_rupture",
+      "hip_dysplasia",
+      "osteoarthritis",
+      "soft_tissue_injury",
+      "impa",
+      "bone_cancer",
+      "ivdd",
+      "degenerative_myelopathy",
+      "iliopsoas_strain",
+      "patellar_luxation",
+      "lumbosacral_disease",
+      "wobbler_syndrome",
+      "obesity_related",
+      "histiocytic_sarcoma",
+    ],
+    follow_up_questions: [
+      "which_leg",
+      "limping_onset",
+      "limping_progression",
+      "weight_bearing",
+      "pain_on_touch",
+      "trauma_history",
+      "worse_after_rest",
+      "swelling_present",
+      "warmth_present",
+      "prior_limping",
+    ],
+    red_flags: ["non_weight_bearing", "visible_fracture", "sudden_paralysis"],
+    body_systems: ["musculoskeletal"],
+  },
+  lethargy: {
+    linked_diseases: [
+      "pain_general",
+      "infection",
+      "anemia",
+      "hypothyroidism",
+      "heart_disease",
+      "kidney_disease",
+      "liver_disease",
+      "addisons_disease",
+      "imha",
+      "babesiosis",
+      "coagulopathy",
+      "sepsis",
+      "heat_stroke",
+      "liver_shunt",
+    ],
+    follow_up_questions: [
+      "lethargy_duration",
+      "lethargy_severity",
+      "appetite_status",
+      "gum_color",
+      "exercise_intolerance",
+    ],
+    red_flags: ["collapse", "unresponsive", "pale_gums"],
+    body_systems: ["systemic"],
+  },
+  coughing: {
+    linked_diseases: [
+      "kennel_cough",
+      "heart_disease",
+      "pneumonia",
+      "collapsing_trachea",
+      "laryngeal_paralysis",
+      "lung_cancer",
+    ],
+    follow_up_questions: [
+      "cough_type",
+      "cough_duration",
+      "cough_timing",
+      "exercise_intolerance",
+      "breathing_rate",
+      "nasal_discharge",
+    ],
+    red_flags: ["breathing_difficulty", "blue_gums", "cough_blood"],
+    body_systems: ["respiratory"],
+  },
+  difficulty_breathing: {
+    linked_diseases: [
+      "pneumonia",
+      "heart_failure",
+      "pleural_effusion",
+      "gdv",
+      "difficulty_breathing",
+      "laryngeal_paralysis",
+      "allergic_reaction",
+      "trauma_chest",
+      "heat_stroke",
+    ],
+    follow_up_questions: [
+      "breathing_onset",
+      "breathing_rate",
+      "gum_color",
+      "position_preference",
+    ],
+    red_flags: ["blue_gums", "pale_gums", "collapse", "breathing_onset_sudden"],
+    body_systems: ["respiratory", "cardiovascular"],
+  },
+  excessive_scratching: {
+    linked_diseases: [
+      "contact_dermatitis",
+      "allergic_dermatitis",
+      "food_allergy",
+      "flea_allergy",
+      "ear_infection",
+      "hot_spots",
+      "mange",
+      "yeast_infection",
+      "zinc_responsive_dermatosis",
+    ],
+    follow_up_questions: [
+      "scratch_location",
+      "scratch_duration",
+      "skin_changes",
+      "flea_prevention",
+      "diet_change",
+      "seasonal_pattern",
+    ],
+    red_flags: ["face_swelling", "hives_widespread"],
+    body_systems: ["dermatologic"],
+  },
+  drinking_more: {
+    linked_diseases: [
+      "diabetes",
+      "cushings_disease",
+      "kidney_disease",
+      "pyometra",
+      "liver_disease",
+      "hypercalcemia",
+    ],
+    follow_up_questions: [
+      "water_amount_change",
+      "urination_frequency",
+      "urination_accidents",
+      "appetite_change",
+      "weight_change",
+      "spay_status",
+    ],
+    red_flags: ["pyometra_signs"],
+    body_systems: ["endocrine", "renal"],
+  },
+  trembling: {
+    linked_diseases: [
+      "situational_stress",
+      "pain_general",
+      "toxin_ingestion",
+      "hypoglycemia",
+      "seizure_disorder",
+      "epilepsy",
+      "eclampsia",
+      "addisons_disease",
+      "fever",
+      "anxiety",
+    ],
+    follow_up_questions: [
+      "trembling_duration",
+      "trembling_timing",
+      "toxin_exposure",
+      "consciousness_level",
+      "restlessness",
+      "temperature_feel",
+    ],
+    red_flags: ["seizure_activity", "toxin_confirmed", "collapse"],
+    body_systems: ["neurologic", "systemic"],
+  },
+  swollen_abdomen: {
+    linked_diseases: [
+      "gdv",
+      "bloat",
+      "ascites",
+      "splenic_mass",
+      "pyometra",
+      "cushings_disease",
+      "pregnancy",
+    ],
+    follow_up_questions: [
+      "abdomen_onset",
+      "abdomen_pain",
+      "unproductive_retching",
+      "spay_status",
+      "restlessness",
+    ],
+    red_flags: ["unproductive_retching", "rapid_onset_distension"],
+    body_systems: ["gastrointestinal", "reproductive"],
+  },
+  blood_in_stool: {
+    linked_diseases: [
+      "hemorrhagic_gastroenteritis",
+      "colitis",
+      "parasites",
+      "foreign_body",
+      "coagulopathy",
+      "gi_cancer",
+      "von_willebrands",
+    ],
+    follow_up_questions: [
+      "blood_color",
+      "blood_amount",
+      "stool_frequency",
+      "toxin_exposure",
+      "rat_poison_access",
+    ],
+    red_flags: ["large_blood_volume", "rat_poison_confirmed", "pale_gums"],
+    body_systems: ["gastrointestinal", "hematologic"],
+  },
+  eye_discharge: {
+    linked_diseases: [
+      "conjunctivitis",
+      "corneal_ulcer",
+      "dry_eye",
+      "glaucoma",
+      "uveitis",
+      "entropion",
+      "cherry_eye",
+      "eye_disorders",
+    ],
+    follow_up_questions: [
+      "discharge_color",
+      "discharge_duration",
+      "squinting",
+      "eye_redness",
+      "vision_changes",
+      "trauma_history",
+    ],
+    red_flags: ["eye_swollen_shut", "eye_bulging", "sudden_blindness"],
+    body_systems: ["ophthalmologic"],
+  },
+  ear_scratching: {
+    linked_diseases: [
+      "ear_infection_bacterial",
+      "ear_infection_yeast",
+      "ear_mites",
+      "allergic_dermatitis",
+      "foreign_body_ear",
+      "aural_hematoma",
+      "syringomyelia",
+    ],
+    follow_up_questions: [
+      "ear_odor",
+      "ear_discharge",
+      "head_shaking",
+      "head_tilt",
+      "balance_issues",
+      "ear_swelling",
+    ],
+    red_flags: ["head_tilt_sudden", "balance_loss", "facial_drooping"],
+    body_systems: ["dermatologic", "neurologic"],
+  },
+  weight_loss: {
+    linked_diseases: [
+      "diabetes",
+      "hyperthyroidism",
+      "kidney_disease",
+      "cancer",
+      "ibd",
+      "exocrine_pancreatic_insufficiency",
+      "parasites",
+      "histiocytic_sarcoma",
+    ],
+    follow_up_questions: [
+      "weight_loss_duration",
+      "weight_loss_amount",
+      "appetite_change",
+      "stool_changes",
+      "water_intake",
+    ],
+    red_flags: ["rapid_weight_loss"],
+    body_systems: ["systemic", "endocrine"],
+  },
+  wound_skin_issue: {
+    linked_diseases: [
+      "wound_infection",
+      "abscess",
+      "hot_spots",
+      "allergic_dermatitis",
+      "skin_mass",
+      "laceration",
+      "autoimmune_skin",
+      "mast_cell_tumor",
+      "perianal_fistula",
+      "alopecia",
+      "zinc_responsive_dermatosis",
+    ],
+    follow_up_questions: [
+      "wound_location",
+      "wound_size",
+      "wound_duration",
+      "wound_color",
+      "wound_discharge",
+      "wound_odor",
+      "wound_licking",
+      "trauma_history",
+    ],
+    red_flags: [
+      "wound_deep_bleeding",
+      "wound_bone_visible",
+      "wound_spreading_rapidly",
+      "wound_tissue_exposed",
+    ],
+    body_systems: ["dermatologic", "musculoskeletal"],
+  },
+  trauma: {
+    linked_diseases: [
+      "soft_tissue_injury",
+      "laceration",
+      "trauma_chest",
+      "pain_general",
+    ],
+    follow_up_questions: [
+      "trauma_mechanism",
+      "trauma_timeframe",
+      "trauma_area",
+      "active_bleeding_trauma",
+      "visible_fracture",
+      "consciousness_level",
+      "gum_color",
+      "breathing_rate",
+      "trauma_mobility",
+    ],
+    red_flags: [
+      "active_bleeding_trauma",
+      "visible_fracture",
+      "blue_gums",
+      "pale_gums",
+      "unresponsive",
+      "inability_to_stand",
+    ],
+    body_systems: ["musculoskeletal", "systemic", "respiratory"],
+  },
+
+  // --- NEW COMPLAINT FAMILIES (VET-902: Dog Complaint Ontology) ---
+
+  seizure_collapse: {
+    linked_diseases: [
+      "seizure_disorder",
+      "epilepsy",
+      "hypoglycemia",
+      "toxin_ingestion",
+      "heart_disease",
+      "imha",
+      "addisons_disease",
+      "heat_stroke",
+      "brain_tumor",
+      "vestibular_disease",
+    ],
+    follow_up_questions: [
+      "seizure_duration",
+      "consciousness_level",
+      "toxin_exposure",
+      "prior_seizures",
+      "trembling_present",
+      "gum_color",
+      "breathing_status",
+    ],
+    red_flags: [
+      "seizure_activity",
+      "collapse",
+      "unresponsive",
+      "toxin_confirmed",
+      "blue_gums",
+      "pale_gums",
+    ],
+    body_systems: ["neurologic", "cardiovascular"],
+  },
+
+  urination_problem: {
+    linked_diseases: [
+      "urinary_stones",
+      "urinary_infection",
+      "prostate_disease",
+      "bladder_cancer",
+      "kidney_disease",
+      "diabetes",
+      "urethral_obstruction",
+    ],
+    follow_up_questions: [
+      "urination_frequency",
+      "straining_present",
+      "blood_in_urine",
+      "urination_accidents",
+      "water_intake",
+      "spay_status",
+    ],
+    red_flags: [
+      "urinary_blockage",
+      "no_urine_24h",
+      "straining_no_urine",
+      "male_unable_to_urinate",
+    ],
+    body_systems: ["renal", "reproductive", "endocrine"],
+  },
+
+  behavior_change: {
+    linked_diseases: [
+      "situational_stress",
+      "cognitive_dysfunction",
+      "brain_tumor",
+      "liver_shunt",
+      "hypothyroidism",
+      "pain_general",
+      "seizure_disorder",
+      "vision_loss",
+      "hearing_loss",
+    ],
+    follow_up_questions: [
+      "behavior_change_duration",
+      "behavior_change_type",
+      "appetite_status",
+      "vision_changes",
+      "sleep_pattern",
+      "recent_events",
+    ],
+    red_flags: ["sudden_disorientation", "new_aggression", "seizure_activity"],
+    body_systems: ["neurologic", "systemic"],
+  },
+
+  swelling_lump: {
+    linked_diseases: [
+      "skin_mass",
+      "mast_cell_tumor",
+      "abscess",
+      "lymphoma",
+      "histiocytic_sarcoma",
+      "allergic_reaction",
+      "lipoma",
+      "hemangiosarcoma",
+    ],
+    follow_up_questions: [
+      "lump_location",
+      "lump_size",
+      "lump_duration",
+      "lump_growth_rate",
+      "lump_mobility",
+      "pain_on_touch",
+      "other_lumps_present",
+    ],
+    red_flags: ["rapid_growing_mass", "face_swelling", "swelling_with_breathing"],
+    body_systems: ["systemic", "dermatologic", "musculoskeletal"],
+  },
+
+  dental_problem: {
+    linked_diseases: [
+      "dental_disease",
+      "oral_tumor",
+      "tooth_root_abscess",
+      "stomatitis",
+      "foreign_body_mouth",
+      "kidney_disease",
+    ],
+    follow_up_questions: [
+      "breath_odor_severity",
+      "drooling_present",
+      "chewing_difficulty",
+      "gum_appearance",
+      "tooth_mobility",
+      "appetite_status",
+    ],
+    red_flags: ["facial_swelling_under_eye", "inability_to_drink", "blood_from_mouth"],
+    body_systems: ["oral", "systemic"],
+  },
+
+  hair_loss: {
+    linked_diseases: [
+      "allergic_dermatitis",
+      "hypothyroidism",
+      "cushings_disease",
+      "mange",
+      "folliculitis",
+      "alopecia",
+      "zinc_responsive_dermatosis",
+      "food_allergy",
+    ],
+    follow_up_questions: [
+      "hair_loss_pattern",
+      "skin_appearance",
+      "itching_present",
+      "hair_loss_duration",
+      "diet_quality",
+      "flea_prevention",
+    ],
+    red_flags: ["widespread_hair_loss_skin_breakdown", "hair_loss_with_lethargy"],
+    body_systems: ["dermatologic", "endocrine"],
+  },
+
+  regurgitation: {
+    linked_diseases: [
+      "megaesophagus",
+      "vascular_ring_anomaly",
+      "myasthenia_gravis",
+      "esophageal_foreign_body",
+      "hiatal_hernia",
+    ],
+    follow_up_questions: [
+      "regurgitation_timing",
+      "food_appearance",
+      "coughing_present",
+      "water_intake",
+      "weight_change",
+      "appetite_status",
+    ],
+    red_flags: ["coughing_after_regurgitation", "blue_gums", "inability_to_keep_water"],
+    body_systems: ["gastrointestinal", "respiratory"],
+  },
+
+  constipation: {
+    linked_diseases: [
+      "obstipation",
+      "prostate_enlargement",
+      "perineal_hernia",
+      "foreign_body",
+      "pelvic_canal_stenosis",
+      "hypothyroidism",
+    ],
+    follow_up_questions: [
+      "last_normal_stool",
+      "straining_duration",
+      "stool_consistency_when_produced",
+      "appetite_status",
+      "vomiting_present",
+      "water_intake",
+    ],
+    red_flags: ["straining_no_production_vomiting", "bloody_rectal_discharge"],
+    body_systems: ["gastrointestinal"],
+  },
+
+  generalized_stiffness: {
+    linked_diseases: [
+      "muscle_soreness",
+      "impa",
+      "osteoarthritis",
+      "polymyositis",
+      "degenerative_myelopathy",
+      "lumbosacral_disease",
+      "hypothyroidism",
+    ],
+    follow_up_questions: [
+      "stiffness_onset",
+      "affected_areas",
+      "fever_present",
+      "appetite_status",
+      "energy_level",
+      "worse_after_rest_or_exercise",
+    ],
+    red_flags: ["fever_with_stiffness", "inability_to_stand", "crying_when_touched"],
+    body_systems: ["musculoskeletal", "systemic"],
+  },
+
+  nasal_discharge: {
+    linked_diseases: [
+      "nasal_infection",
+      "nasal_tumor",
+      "nasal_foreign_body",
+      "aspergillosis",
+      "dental_disease",
+      "allergic_rhinitis",
+    ],
+    follow_up_questions: [
+      "discharge_color",
+      "discharge_side",
+      "sneezing_frequency",
+      "blood_present",
+      "appetite_status",
+      "nasal_discharge_duration",
+    ],
+    red_flags: ["bloody_nasal_discharge_one_sided", "facial_deformity"],
+    body_systems: ["respiratory"],
+  },
+
+  vaginal_discharge: {
+    linked_diseases: [
+      "pyometra",
+      "vaginal_hyperplasia",
+      "metritis",
+      "vaginal_tumor",
+      "urinary_infection",
+    ],
+    follow_up_questions: [
+      "spay_status",
+      "vaginal_discharge_color",
+      "discharge_odor",
+      "heat_cycle_timing",
+      "appetite_status",
+      "water_intake",
+    ],
+    red_flags: ["intact_female_lethargy_drinking", "foul_smelling_discharge", "collapse"],
+    body_systems: ["reproductive"],
+  },
+
+  testicular_prostate: {
+    linked_diseases: [
+      "prostate_disease",
+      "testicular_tumor",
+      "prostatitis",
+      "perineal_hernia",
+      "benign_prostatic_hyperplasia",
+    ],
+    follow_up_questions: [
+      "neuter_status",
+      "swelling_location",
+      "urination_changes",
+      "prostate_stool_changes",
+      "pain_on_touch",
+      "testicular_prostate_duration",
+    ],
+    red_flags: ["acute_painful_testicular_swelling", "inability_to_urinate"],
+    body_systems: ["reproductive", "renal"],
+  },
+
+  exercise_induced_lameness: {
+    linked_diseases: [
+      "muscle_soreness",
+      "ccl_rupture",
+      "iliopsoas_strain",
+      "heart_disease",
+      "exercise_induced_collapse",
+      "myopathy",
+      "osteoarthritis",
+    ],
+    follow_up_questions: [
+      "exercise_type",
+      "onset_during_exercise",
+      "recovery_time",
+      "breathing_after_exercise",
+      "gum_color",
+      "prior_episodes",
+    ],
+    red_flags: ["collapse_after_exercise", "blue_gums_after_exercise"],
+    body_systems: ["musculoskeletal", "cardiovascular"],
+  },
+
+  skin_odor_greasy: {
+    linked_diseases: [
+      "yeast_infection",
+      "seborrhea",
+      "allergic_dermatitis",
+      "hypothyroidism",
+      "cushings_disease",
+    ],
+    follow_up_questions: [
+      "odor_location",
+      "skin_appearance",
+      "itching_present",
+      "bath_frequency",
+      "ear_involvement",
+      "diet_quality",
+    ],
+    red_flags: ["widespread_skin_breakdown", "fever_with_skin_odor"],
+    body_systems: ["dermatologic"],
+  },
+
+  recurrent_ear: {
+    linked_diseases: [
+      "allergic_dermatitis",
+      "ear_infection_bacterial",
+      "ear_infection_yeast",
+      "food_allergy",
+      "hypothyroidism",
+    ],
+    follow_up_questions: [
+      "infection_frequency",
+      "last_treatment",
+      "underlying_allergy_diagnosis",
+      "food_trial_done",
+      "ear_cleaning_routine",
+    ],
+    red_flags: ["head_tilt_sudden", "balance_loss", "facial_drooping"],
+    body_systems: ["dermatologic", "systemic"],
+  },
+
+  recurrent_skin: {
+    linked_diseases: [
+      "allergic_dermatitis",
+      "superficial_pyoderma",
+      "demodicosis",
+      "hypothyroidism",
+      "cushings_disease",
+      "immune_deficiency",
+    ],
+    follow_up_questions: [
+      "skin_infection_frequency",
+      "antibiotic_history",
+      "allergy_testing_done",
+      "immune_status",
+      "diet_quality",
+    ],
+    red_flags: ["widespread_deep_infections", "fever", "non_responsive_to_antibiotics"],
+    body_systems: ["dermatologic", "systemic", "immune"],
+  },
+
+  inappropriate_urination: {
+    linked_diseases: [
+      "urinary_infection",
+      "urinary_stones",
+      "diabetes",
+      "cushings_disease",
+      "kidney_disease",
+      "prostate_disease",
+      "cognitive_dysfunction",
+    ],
+    follow_up_questions: [
+      "urination_frequency",
+      "straining_present",
+      "blood_in_urine",
+      "water_intake",
+      "neuter_status",
+      "behavioral_changes",
+    ],
+    red_flags: ["straining_no_urine", "blood_in_urine", "male_unable_to_urinate"],
+    body_systems: ["renal", "reproductive", "behavioral"],
+  },
+
+  fecal_incontinence: {
+    linked_diseases: [
+      "ivdd",
+      "lumbosacral_disease",
+      "cauda_equina_syndrome",
+      "anal_sphincter_incompetence",
+      "cognitive_dysfunction",
+    ],
+    follow_up_questions: [
+      "fecal_incontinence_onset",
+      "stool_consistency",
+      "hind_limb_function",
+      "tail_movement",
+      "back_pain",
+      "perineal_reflex",
+    ],
+    red_flags: ["sudden_onset_hind_weakness", "tail_paralysis", "inability_to_stand"],
+    body_systems: ["gastrointestinal", "neurologic"],
+  },
+
+  vomiting_diarrhea_combined: {
+    linked_diseases: [
+      "dietary_indiscretion",
+      "gastroenteritis",
+      "pancreatitis",
+      "parvovirus",
+      "toxin_ingestion",
+      "foreign_body",
+      "hemorrhagic_gastroenteritis",
+    ],
+    follow_up_questions: [
+      "combined_vomiting_duration",
+      "combined_diarrhea_duration",
+      "blood_in_either",
+      "appetite_status",
+      "water_intake",
+      "gum_color",
+      "vaccination_status",
+      "toxin_exposure",
+    ],
+    red_flags: ["not_drinking", "large_blood_volume", "pale_gums"],
+    body_systems: ["gastrointestinal"],
+  },
+
+  coughing_breathing_combined: {
+    linked_diseases: [
+      "heart_failure",
+      "pneumonia",
+      "pleural_effusion",
+      "laryngeal_paralysis",
+      "allergic_reaction",
+    ],
+    follow_up_questions: [
+      "breathing_rate",
+      "gum_color",
+      "cough_type",
+      "coughing_breathing_onset",
+      "exercise_intolerance",
+      "position_preference",
+    ],
+    red_flags: ["blue_gums", "collapse", "sudden_onset", "inability_to_lie_down"],
+    body_systems: ["respiratory", "cardiovascular"],
+  },
+
+  oral_mass: {
+    linked_diseases: [
+      "oral_tumor",
+      "epulis",
+      "melanoma",
+      "squamous_cell_carcinoma",
+      "foreign_body_mouth",
+      "dental_disease",
+    ],
+    follow_up_questions: [
+      "oral_mass_location",
+      "oral_mass_size",
+      "bleeding_present",
+      "eating_difficulty",
+      "oral_mass_duration",
+      "breath_odor_severity",
+    ],
+    red_flags: ["bleeding_from_mouth", "inability_to_eat_drink", "facial_swelling"],
+    body_systems: ["oral"],
+  },
+
+  vision_loss: {
+    linked_diseases: [
+      "sudden_acquired_retinal_degeneration",
+      "glaucoma",
+      "cataract",
+      "optic_neuritis",
+      "brain_tumor",
+      "hypertension",
+    ],
+    follow_up_questions: [
+      "vision_loss_onset",
+      "one_or_both_eyes",
+      "pain_present",
+      "pupil_appearance",
+      "other_neurologic_signs",
+      "vision_loss_duration",
+    ],
+    red_flags: ["sudden_blindness", "painful_eye", "dilated_non_responsive_pupils"],
+    body_systems: ["ophthalmologic", "neurologic"],
+  },
+
+  hearing_loss: {
+    linked_diseases: [
+      "ear_infection",
+      "vestibular_disease",
+      "age_related_deafness",
+      "ototoxicity",
+      "brain_tumor",
+    ],
+    follow_up_questions: [
+      "hearing_loss_onset",
+      "ear_infection_history",
+      "head_tilt",
+      "balance_issues",
+      "response_to_loud_sounds",
+      "dog_age_years",
+    ],
+    red_flags: ["sudden_deafness_head_tilt", "balance_loss", "facial_drooping"],
+    body_systems: ["neurologic", "dermatologic"],
+  },
+
+  aggression: {
+    linked_diseases: [
+      "pain_general",
+      "ivdd",
+      "dental_disease",
+      "ear_infection",
+      "cognitive_dysfunction",
+      "hypothyroidism",
+      "brain_tumor",
+    ],
+    follow_up_questions: [
+      "aggression_onset",
+      "trigger_situations",
+      "pain_on_touch",
+      "appetite_status",
+      "energy_level",
+      "recent_events",
+    ],
+    red_flags: ["sudden_new_aggression", "aggression_with_trembling", "aggression_with_lethargy"],
+    body_systems: ["systemic", "musculoskeletal", "neurologic"],
+  },
+
+  pacing_restlessness: {
+    linked_diseases: [
+      "gdv",
+      "pain_general",
+      "bloat",
+      "cognitive_dysfunction",
+      "situational_stress",
+      "anxiety",
+      "splenic_mass",
+    ],
+    follow_up_questions: [
+      "abdomen_appearance",
+      "retching_present",
+      "gum_color",
+      "pacing_duration",
+      "appetite_status",
+      "water_intake",
+    ],
+    red_flags: ["pacing_with_bloated_abdomen", "pacing_with_retching", "pacing_over_2h"],
+    body_systems: ["systemic", "neurologic", "gastrointestinal"],
+  },
+
+  abnormal_gait: {
+    linked_diseases: [
+      "ivdd",
+      "degenerative_myelopathy",
+      "wobbler_syndrome",
+      "vestibular_disease",
+      "fibrocartilaginous_embolism",
+      "brain_tumor",
+    ],
+    follow_up_questions: [
+      "abnormal_gait_onset",
+      "affected_limbs",
+      "back_pain",
+      "bladder_control",
+      "trauma_history",
+      "abnormal_gait_progression",
+    ],
+    red_flags: ["inability_to_stand", "paralysis", "loss_bladder_bowel_control"],
+    body_systems: ["neurologic", "musculoskeletal"],
+  },
+
+  heat_intolerance: {
+    linked_diseases: [
+      "heat_stroke",
+      "difficulty_breathing",
+      "heart_disease",
+      "obesity_related",
+    ],
+    follow_up_questions: [
+      "temperature_exposure",
+      "heat_exposure_duration",
+      "gum_color",
+      "consciousness_level",
+      "vomiting_present",
+      "water_intake",
+    ],
+    red_flags: ["collapse_in_heat", "brick_red_gums", "vomiting_overheating"],
+    body_systems: ["respiratory", "systemic"],
+  },
+
+  postoperative_concern: {
+    linked_diseases: [
+      "wound_infection",
+      "surgical_complication",
+      "seroma",
+      "dehiscence",
+      "pain_general",
+    ],
+    follow_up_questions: [
+      "surgery_type",
+      "days_post_op",
+      "incision_appearance",
+      "discharge_present",
+      "appetite_status",
+      "activity_level",
+    ],
+    red_flags: ["incision_dehiscence", "active_bleeding_incision", "pus_with_fever"],
+    body_systems: ["systemic", "dermatologic"],
+  },
+
+  medication_reaction: {
+    linked_diseases: [
+      "allergic_reaction",
+      "toxin_ingestion",
+      "gastroenteritis",
+    ],
+    follow_up_questions: [
+      "medication_name",
+      "medication_dose",
+      "medication_timing",
+      "reaction_symptoms",
+      "prior_reactions",
+      "current_medications",
+    ],
+    red_flags: ["face_swelling", "hives_with_breathing", "collapse"],
+    body_systems: ["systemic"],
+  },
+  post_vaccination_reaction: {
+    linked_diseases: [
+      "allergic_reaction",
+      "gastroenteritis",
+      "pain_general",
+      "fever",
+    ],
+    follow_up_questions: [
+      "vaccination_timing",
+      "vaccination_type",
+      "reaction_symptoms",
+      "face_swelling",
+      "hives_with_breathing",
+      "fever_present",
+      "appetite_status",
+    ],
+    red_flags: ["face_swelling", "hives_with_breathing", "unresponsive"],
+    body_systems: ["systemic", "dermatologic"],
+  },
+
+  pregnancy_birth: {
+    linked_diseases: [
+      "dystocia",
+      "metritis",
+      "eclampsia",
+      "pregnancy",
+      "pyometra",
+    ],
+    follow_up_questions: [
+      "days_pregnant",
+      "contraction_status",
+      "discharge_color",
+      "puppies_delivered",
+      "time_since_last_puppy",
+      "restlessness",
+      "appetite_status",
+    ],
+    red_flags: ["dystocia_active", "green_discharge_no_puppy", "eclampsia_signs"],
+    body_systems: ["reproductive"],
+  },
+
+  puppy_concern: {
+    linked_diseases: [
+      "hypoglycemia",
+      "parasites",
+      "parvovirus",
+      "fading_puppy_syndrome",
+      "congenital_defect",
+      "liver_shunt",
+    ],
+    follow_up_questions: [
+      "puppy_age_weeks",
+      "nursing_status",
+      "puppy_temperature",
+      "weight_trend",
+      "littermate_status",
+      "vaccination_status",
+    ],
+    red_flags: ["puppy_not_nursing_4h", "puppy_cold_to_touch", "diarrhea_under_12_weeks"],
+    body_systems: ["systemic"],
+  },
+
+  senior_decline: {
+    linked_diseases: [
+      "cognitive_dysfunction",
+      "osteoarthritis",
+      "kidney_disease",
+      "heart_disease",
+      "cancer",
+      "hypothyroidism",
+      "dental_disease",
+    ],
+    follow_up_questions: [
+      "senior_decline_duration",
+      "specific_changes",
+      "appetite_status",
+      "water_intake",
+      "mobility_level",
+      "sleep_pattern",
+    ],
+    red_flags: ["rapid_decline_weeks", "inability_to_stand", "sudden_blindness"],
+    body_systems: ["neurologic", "systemic", "musculoskeletal"],
+  },
+
+  multi_system_decline: {
+    linked_diseases: [
+      "kidney_disease",
+      "liver_disease",
+      "cancer",
+      "addisons_disease",
+      "imha",
+      "heart_failure",
+      "sepsis",
+    ],
+    follow_up_questions: [
+      "each_symptom_duration",
+      "appetite_status",
+      "water_intake",
+      "weight_change",
+      "energy_level",
+      "vomiting_present",
+    ],
+    red_flags: ["lethargy_not_eating_not_drinking", "pale_gums_collapse"],
+    body_systems: ["systemic"],
+  },
+
+  unknown_concern: {
+    linked_diseases: ["pain_general", "infection"],
+    follow_up_questions: [
+      "chief_complaint_guess",
+      "appetite_status",
+      "water_intake",
+      "energy_level",
+      "breathing_status",
+      "gum_color",
+      "last_normal",
+    ],
+    red_flags: ["unable_to_assess_breathing", "dog_non_responsive"],
+    body_systems: ["unknown"],
+  },
+};
+
